% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{add_pkgs_installation_code}
\alias{add_pkgs_installation_code}
\title{Add column with package instalation code.}
\usage{
add_pkgs_installation_code(pkgs_df)
}
\arguments{
\item{pkgs_df}{Data frame with columns "paketas" (character) and "on_cran" (logical)}
}
\value{
The same data frame with additional columns
\itemize{
\item "diegimo_kodas"  -- strings with code; if a package is not on cran and its
installation code is not provided, NA value is returned.
\item "ar_mod_kodas" -- logical, if modified code for instalation should be used.
}
}
\description{
Add column with installation code for each provided package.
}
