% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio-package.R
\docType{package}
\name{bio-package}
\alias{bio}
\alias{bio-package}
\title{bio: Resources Manager for R, RStudio and Related Software}
\description{
Provides tools for managing R and RStudio resources in Biostatistics teaching environments. It configures user preferences, keybindings, spellcheck dictionaries, project lists, and course setup, and checks installed software and available versions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mokymai.github.io/bio/}
  \item \url{https://github.com/mokymai/bio/}
  \item Report bugs at \url{https://github.com/mokymai/bio/issues}
}

}
\author{
\strong{Maintainer}: Vilmantas Gegzna \email{GegznaV@gmail.com} (\href{https://orcid.org/0000-0002-9500-5167}{ORCID})

Authors:
\itemize{
  \item Vilmantas Gegzna \email{GegznaV@gmail.com} (\href{https://orcid.org/0000-0002-9500-5167}{ORCID})
}

}
\keyword{internal}
