% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio-related.R
\name{bio_version}
\alias{bio_version}
\alias{get_vesion_bio}
\alias{check_updates_bio}
\alias{update_pkg_bio}
\title{Version management of package 'bio'.}
\usage{
get_vesion_bio()

check_updates_bio(show_status = "always", install = "outdated", ...)

update_pkg_bio()
}
\description{
\code{get_vesion_bio()} returns version number of package \pkg{bio}.
\code{check_updates_bio()} checks if package \pkg{bio} has a recommended version.
\code{update_bio()} tries to update package \pkg{bio}.
}
\examples{
get_vesion_bio()

\dontrun{\donttest{

check_updates_bio()
update_bio()

}}
}
