% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.impose_deadtime}
\alias{bursts.impose_deadtime}
\title{Imposes a deadtime to a burst by applying segment.impose_deadtime to each segment in a burst}
\usage{
bursts.impose_deadtime(burst, dead_time)
}
\arguments{
\item{burst, }{a burst containing segments of dwells and states.}

\item{dead_time, }{the briefest possible event in microseconds.}
}
\value{
A modified copy of the original burst
}
\description{
The user specifies a deadtime in microseconds. The function applies
segment.impose_deadtime to each segment in the burst.
(See segment.impose_deadtime for details.)
}
