% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bursts.R
\name{bursts.subconductance_as}
\alias{bursts.subconductance_as}
\title{Imposes a fixed conductance level (0 or 1) to all dwells with subconductance levels to each segment in a burst by applying segment.subconductance_as.}
\usage{
bursts.subconductance_as(burst, level)
}
\arguments{
\item{burst, }{a burst containing segments of dwells and states.}

\item{level, }{either 'open' or 'closed'}
}
\value{
A modified copy of the original burst
}
\description{
The user specifies the desired level ('open' or 'closed'). The function applies
segment.subconductance_as to each segment in the burst.
(See segment.subconductance_as for details.)
}
