% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{check_installed_packages}
\alias{check_installed_packages}
\title{Check installed packages}
\usage{
check_installed_packages(
  list_name,
  use_local_list = getOption("bio.use_local_list", FALSE),
  ...
)
}
\arguments{
\item{list_name}{(character) The name of the list with required R packages.
E.g., "mini", "Rcmdr", etc.}

\item{use_local_list}{(logical) If \code{TRUE}, the list, which is in the folder
of package \pkg{bio} ("local list"), is used. If \code{FALSE}, the lists
on "GitHub" repository of the package is used. It is recommended
using the online version of the list, as it may contain more recent
changes.}

\item{...}{Further arguments to \code{\link[=get_pkgs_installation_status]{get_pkgs_installation_status()}}.}
}
\value{
Function invisibly returns object with package installation status.
}
\description{
A user-fiendly version of a function to check if required R packages are
installed and have minimum required versions.
}
\examples{
\dontrun{\donttest{
check_installed_packages("mini", use_local_list = TRUE)

check_installed_packages("mini", include = "always", use_local_list = TRUE)
check_installed_packages("mini", include = "always", install = "outdated", github = "always", use_local_list = TRUE)
}}
}
