% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages--check.R
\name{check_packages}
\alias{check_packages}
\alias{check_packages_by_topic}
\alias{check_packages_by_name}
\title{Check Installed Packages by Topic and by Name}
\usage{
check_packages_by_topic(
  list_name = NULL,
  use_local_list = getOption("bio.use_local_list", FALSE),
  upgrade = TRUE,
  ...
)

check_packages_by_name(
  pkgs = NULL,
  use_local_list = getOption("bio.use_local_list", FALSE),
  upgrade = TRUE,
  ...
)
}
\arguments{
\item{list_name}{(character) The name of the list with required R packages.
E.g., "mini", "Rcmdr", "Rcmdr-biostat", "bio", etc.}

\item{use_local_list}{(logical) If \code{TRUE}, the list, which is locally
installed in the folder of package \pkg{bio} ("local list"), is used.
If \code{FALSE}, the list on "GitHub" repository of the package is used.
It is recommended  using the online version of the list, as it may
contain more recent changes.
Optiom bight be set globally by, e.g.,
\code{options(bio.use_local_list = TRUE)}.}

\item{upgrade}{Upgrade dependencies.
See \code{upgrade} in \code{\link[remotes:install_cran]{remotes::install_cran()}}.}

\item{...}{Further arguments to \code{\link[=get_pkgs_installation_status]{get_pkgs_installation_status()}}.}

\item{pkgs}{(character) The name (vector of names) of R packages.
Either \code{list_name} or \code{pkgs} must be \code{NULL}.}
}
\value{
Function invisibly returns object with package installation status.
}
\description{
A user-friendly version of functions to check if required R packages are
installed and have minimum required versions.
}
\examples{
if (FALSE) {
  check_packages_by_topic("mini", use_local_list = TRUE)

  check_packages_by_topic("mini", include = "always", use_local_list = TRUE)

  check_packages_by_topic("mini", include = "always", install = "outdated",
   github = "always", use_local_list = TRUE)


  check_packages_by_name("bio")

  check_packages_by_name(c("bio", "usethis", "ggplot1"))
}
}
\concept{check}
\concept{check-packages}
\concept{packages}
