% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-functions.R
\name{check_packages_by_topic-defunct}
\alias{check_packages_by_topic-defunct}
\alias{check_installed_packages}
\title{DEFUNCT. Use \code{check_packages_by_topic()} instead}
\usage{
check_installed_packages(...)
}
\arguments{
\item{...}{See \code{\link[=check_packages_by_topic]{check_packages_by_topic()}}.}
}
\description{
\strong{DO NOT USE THIS FUNCTION!}
This function was replaced by  \code{\link[=check_packages_by_topic]{check_packages_by_topic()}} and will be
removed soon.
}
\concept{defunct}
