% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{check_updates_bio}
\alias{check_updates_bio}
\title{Deprecated. Use \code{check_updates_pkg_bio()} instead}
\usage{
check_updates_bio(
  show_status = "always",
  install = "outdated",
  upgrade = FALSE,
  ...
)
}
\description{
This function was replaced by \code{\link[=check_updates_pkg_bio]{check_updates_pkg_bio()}} and will be removed
soon.
}
\concept{deprecated}
