% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{check_user_info}
\alias{check_user_info}
\title{Get user-related information}
\usage{
check_user_info()
}
\description{
Prints user and system metadata useful for troubleshooting local R setups.
}
\examples{
if (interactive()) {
  check_user_info()
}
}
\concept{check}
\concept{check-user-info}
