% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{clear_rs_history}
\alias{clear_rs_history}
\title{DEPRECATED. Use \code{rstudio_clear_history()} instead}
\usage{
clear_rs_history(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_clear_history]{rstudio_clear_history()}}}
}
\description{
DO NOT USE THIS FUNCTION!
This function was replaced by \code{\link[=rstudio_clear_history]{rstudio_clear_history()}} and will be removed soon.
}
\concept{deprecated}
