% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages--check.R
\name{compare_version}
\alias{compare_version}
\title{Compare Version Numbers}
\usage{
compare_version(v_installed, v_required)
}
\arguments{
\item{v_installed}{vector with installed version numbers}

\item{v_required}{vector with required version numbers}
}
\value{
The same as in \code{\link[utils:compareVersion]{utils::compareVersion()}}, just a vector.
}
\description{
Compare Version Numbers
}
\examples{

compare_version("2.4", "2")

compare_version("2.3", "2.3")

compare_version("2.3", "2.3.1")

}
\seealso{
Other R-packages-related functions:
\code{\link[=get_pkgs_installed]{get_pkgs_installed()}}
}
\concept{R-packages-related functions}
\concept{utilities}
