% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_hjcfit_likelihood}
\alias{cpp_hjcfit_likelihood}
\title{Computes the likelihood of a kinetic model given observed bursts,
with missed event correction, using the HJCFIT library. This
function is an interface for Log10Likelihood from HJCFIT.}
\usage{
cpp_hjcfit_likelihood(qmatrix, nopen, bursts, tau, nmax = 2L,
  xtol = 1e-10, rtol = 1e-10, itermax = 100L)
}
\arguments{
\item{qmatrix}{Initial guess qmatrix.}

\item{nopen}{Number of open states in the matrix}

\item{bursts}{A list of bursts}

\item{tau}{Maximum length of the missed events}

\item{nmax}{The exact missed-event likelihood will be computed for 't < n_max * tau'}

\item{xtol}{Tolerance criteria for brentq().}

\item{rtol}{Tolerance criteria for brentq().}

\item{itermax}{Maximum number of iteration when calling brentq().}
}
\value{
The negative log10-likelihood of the model given the burst sequence.
}
\description{
https://github.com/DCPROGS/HJCFIT
}
\details{
See the link above and the references therein for details.

NOTE: It is slow to call this over and over again.
}
