% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_hjcfit_likelihood_maximize}
\alias{cpp_hjcfit_likelihood_maximize}
\title{A maximum likelihood estimate for a Q-Matrix with a fixed number
of states. Computes the likelihood of a kinetic model given
observed bursts, with missed event correction, using the HJCFIT
library. This function is an interface for Log10Likelihood from
HJCFIT.}
\usage{
cpp_hjcfit_likelihood_maximize(qmatrix, nopen, bursts, tau, nmax = 2L,
  xtol = 1e-10, rtol = 1e-10, itermax = 100L, REQMIN = 1e-06,
  KONVGE = 10L, KMAX = 1000L)
}
\arguments{
\item{qmatrix}{Initial guess qmatrix.}

\item{nopen}{Number of open states in the matrix.}

\item{bursts}{A list of bursts.}

\item{tau}{Maximum length of the missed events.}

\item{nmax}{The exact missed-event likelihood will be computed for 't < n_max * tau'}

\item{xtol}{Tolerance criteria for brentq().}

\item{rtol}{Tolerance criteria for brentq().}

\item{itermax}{Maximum number of iteration when calling brentq().}

\item{REQMIN}{Nelder-Mead parameter: stop if within required distance.}

\item{KONVGE}{Nelder-Mead parameter: check convergence every KONVGE steps.}

\item{KMAX}{Nelder-Mead parameter: maximum number of iterations.}
}
\value{
A qmatrix kinetic model which numerically maximizes likelihood.
}
\description{
https://github.com/DCPROGS/HJCFIT
}
\details{
See the link above and the references therein for details.
}
