% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{deprecated-functions}
\alias{deprecated-functions}
\alias{set_rstudio_keybindings}
\alias{update_bio}
\alias{restart_rstudio}
\title{Deprecated functions}
\usage{
set_rstudio_keybindings(...)

update_bio(...)

restart_rstudio()
}
\arguments{
\item{...}{Arguments to further methods.}
}
\description{
Functions that are no longer used in package \pkg{bio}.
In the next release of the package, these functions will be removed.
}
