% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_task_names.R
\name{emokymai}
\alias{emokymai}
\alias{emokymai_browse_assingment}
\alias{emokymai_submit_assingment}
\title{Open Resource on Emokymai}
\usage{
emokymai_browse_assingment(id)

emokymai_submit_assingment(id)
}
\arguments{
\item{id}{Resource ID.}
}
\description{
Open Resource on Emokymai
}
