% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{extract_proj_name}
\alias{extract_proj_name}
\title{Extract R Project's Name}
\usage{
extract_proj_name(path)
}
\arguments{
\item{path}{(character) Vector of character names}
}
\value{
(CHaracter) a vector of extracted character names
}
\description{
Extract R Project's Name
}
\examples{
path_to_project <- c("list/proj.Rproj", "C:/R/BS-2020/bs.Rproj")
extract_proj_name(path_to_project)
}
