% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings--preferences.R
\name{format_pref_value}
\alias{format_pref_value}
\title{Format a preference value for display in a one-line diff summary.}
\usage{
format_pref_value(x)
}
\description{
Format a preference value for display in a one-line diff summary.
}
\keyword{internal}
