% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages--find.R
\name{get_base_packages}
\alias{get_base_packages}
\title{Get Standard Base R Packages}
\usage{
get_base_packages()
}
\value{
Character vector of base R package names.
}
\description{
Returns a character vector of standard packages included in the base R
distribution to allow filtering them out from dependencies.
}
\keyword{internal}
