% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_modified_pkgs_installation_code}
\alias{get_modified_pkgs_installation_code}
\title{Get modified package instalation code.}
\usage{
get_modified_pkgs_installation_code(file = "pkgs-install-from")
}
\value{
Dataframe with columns  "paketas" and "diegimo_kodas".
}
\description{
Get installation code of packages that either should be installed not from
CRAN or a modified code shold be installed.
}
