% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os--NEW.R, R/get_os.R
\name{get_os_type}
\alias{get_os_type}
\alias{is_64bit_os}
\alias{is_32bit_os}
\title{Determine Operating System (OS)}
\usage{
get_os_type()

is_64bit_os()

is_32bit_os()

get_os_type()

is_32bit_os()

is_64bit_os()
}
\value{
OS name in lower case: "windows", "mac", "linux", "sunos", etc.

OS name in lower case: windows, mac, linux, etc.
}
\description{
Determine the operating system of the current R session.

Determine the operating system (OS) of your machine.
}
\examples{
get_os_type()
get_os_type()
}
\seealso{
https://www.r-bloggers.com/identifying-the-os-from-r/
}
\author{
The code is based on https://www.r-bloggers.com/identifying-the-os-from-r/
}
\concept{utilities}
