% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{get_path_desktop}
\alias{get_path_desktop}
\alias{open_desktop}
\title{Path to Desktop Folder}
\usage{
get_path_desktop(...)

open_desktop()
}
\arguments{
\item{...}{(character) file or folder name on desktop.}
}
\value{
String with path to desktop or path to file or folder on a desktop.
}
\description{
Get path to desktop folder of current user and open it.
}
\examples{
get_path_desktop()
}
\concept{paths and dirs}
