% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{get_path_to_git}
\alias{get_path_to_git}
\title{Get path to Git.}
\usage{
get_path_to_git()
}
\value{
Path to Git as string of NULL, if Git is not installed or not configured properly.
}
\description{
Get path to Git.
}
\examples{
get_path_to_git()
}
