% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_cran_details}
\alias{get_pkgs_cran_details}
\title{Get details about package on CRAN}
\usage{
get_pkgs_cran_details()
}
\value{
Data frame with columns "package", "cran_version", "on_cran".
}
\description{
Convenience function based on \code{\link[utils:available.packages]{utils::available.packages()}}.
}
\examples{
\dontrun{\donttest{

# NOTE: Internet connection is needed.
head(get_pkgs_cran_details())

}}
}
\seealso{
\code{\link[utils:available.packages]{utils::available.packages()}}

Other R-packages-related functions: 
\code{\link{compare_version}()},
\code{\link{get_pkgs_installation_status_local}()},
\code{\link{get_pkgs_installed}()},
\code{\link{get_pkgs_non_cran_installation_details}()},
\code{\link{get_pkgs_recommended}()},
\code{\link{get_pkgs_req_version}()}
}
\concept{R-packages-related functions}
\concept{packages}
