% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages--check.R
\name{get_pkgs_installed}
\alias{get_pkgs_installed}
\title{List packages installed on this computer}
\usage{
get_pkgs_installed(rm_duplicates = TRUE)
}
\arguments{
\item{rm_duplicates}{(logical) Should duplicated names of packages be removed?
If \code{TRUE}, when several packages are found, only the one with the
highest version is returned. If \code{FALSE}, no packages are removed
from the list.}
}
\value{
Data frame with columns \code{"package"} and \code{"current_version"}.
}
\description{
List packages installed on this computer
}
\examples{

head(get_pkgs_installed())

nrow(get_pkgs_installed(rm_duplicates = TRUE))
nrow(get_pkgs_installed(rm_duplicates = FALSE))

}
\seealso{
Other R-packages-related functions:
\code{\link[=compare_version]{compare_version()}},
\code{\link[=get_pkgs_installation_status]{get_pkgs_installation_status()}}
}
\concept{R-packages-related functions}
\concept{packages}
