% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_non_cran_installation_details}
\alias{get_pkgs_non_cran_installation_details}
\title{Get details about non-CRAN package installation.}
\usage{
get_pkgs_non_cran_installation_details(
  use_local_list = getOption("bio.use_local_list", FALSE)
)
}
\arguments{
\item{use_local_list}{(logical) If \code{TRUE}, the list, which is locally
installed in the folder of package \pkg{bio} ("local list"), is used.
If \code{FALSE}, the list on "GitHub" repository of the package is used.
It is recommended  using the online version of the list, as it may
contain more recent changes.
Optiom bight be set globally by, e.g.,
\code{options(bio.use_local_list = TRUE)}.}
}
\value{
Dataframe with columns:
\itemize{
\item \code{"package"} package name,
\item \code{"install_from"} source to install from. Currently supported values
are "code" and "github"
\item \code{"details"} either installation code of repository name
\item \code{"notes"}  notes for user.
}
}
\description{
Get installation code of packages that either should be installed not from
CRAN or a modified code shold be installed.
}
\details{
If the file has column \code{ignore} with value \code{TRUE}, the line is not included
in the output.
}
\examples{
# NOTE: It is not recommended to use the local lists as they might be out of date.
options(bio.use_local_list = TRUE)

head(get_pkgs_non_cran_installation_details())
}
\seealso{
Other R-packages-related functions: 
\code{\link{compare_version}()},
\code{\link{get_pkgs_cran_details}()},
\code{\link{get_pkgs_installation_status_local}()},
\code{\link{get_pkgs_installed}()},
\code{\link{get_pkgs_recommended}()},
\code{\link{get_pkgs_req_version}()}
}
\concept{R-packages-related functions}
