% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_recommended}
\alias{get_pkgs_recommended}
\title{List of packages of interest.}
\usage{
get_pkgs_recommended(file = "install-r/pkgs-recommended.txt")
}
\value{
Data frame with column \code{"package"}.
}
\description{
List of packages of interest.
}
\examples{
\dontrun{\donttest{

head(get_pkgs_recommended())

}}
}
\seealso{
Other R-packages-related functions: 
\code{\link{compare_version}()},
\code{\link{get_pkgs_cran_details}()},
\code{\link{get_pkgs_instalation_status}()},
\code{\link{get_pkgs_installed}()},
\code{\link{get_pkgs_non_cran_installation_details}()},
\code{\link{get_pkgs_req_version}()}
}
\concept{R-packages-related functions}
