% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_req_version}
\alias{get_pkgs_req_version}
\title{Get required version of packages}
\usage{
get_pkgs_req_version(use_local_list = getOption("bio.use_local_list", FALSE))
}
\arguments{
\item{use_local_list}{(logical) If \code{TRUE}, the list, which is locally
installed in the folder of package \pkg{bio} ("local list"), is used.
If \code{FALSE}, the list on "GitHub" repository of the package is used.
It is recommended  using the online version of the list, as it may
contain more recent changes.
Optiom bight be set globally by, e.g.,
\code{options(bio.use_local_list = TRUE)}.}
}
\value{
Dataframe with columns  "package" and "required_version".
}
\description{
Get required version of packages from a list in a file.
}
\examples{
# NOTE: It is not recommended to use the local lists as they might be out of date.
options(bio.use_local_list = TRUE)

head(get_pkgs_req_version())

}
\seealso{
Other R-packages-related functions: 
\code{\link{compare_version}()},
\code{\link{get_pkgs_cran_details}()},
\code{\link{get_pkgs_installation_status_local}()},
\code{\link{get_pkgs_installed}()},
\code{\link{get_pkgs_non_cran_installation_details}()},
\code{\link{get_pkgs_recommended}()}
}
\concept{R-packages-related functions}
\concept{packages}
