% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{get_rs_file_ids}
\alias{get_rs_file_ids}
\alias{get_rs_file_ids_user}
\alias{get_rs_file_ids_project}
\title{Get RStudio document IDs and paths}
\usage{
get_rs_file_ids()

get_rs_file_ids_user()

get_rs_file_ids_project()
}
\value{
Data frame with columns \code{id} (for IDs) and \code{path} (for file paths).
}
\description{
Get RStudio file IDs and paths.
}
\examples{
\dontrun{\donttest{
head(get_rs_file_ids_user())
}}
}
\concept{paths and dirs}
