% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{get_rs_ui_prefs}
\alias{get_rs_ui_prefs}
\title{DEPRECATED. Use \code{rstudio_get_ui_prefs()} instead}
\usage{
get_rs_ui_prefs(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_get_ui_prefs]{rstudio_get_ui_prefs()}}}
}
\description{
\strong{DO NOT USE THIS FUNCTION!}
This function was replaced by \code{\link[=rstudio_get_ui_prefs]{rstudio_get_ui_prefs()}} and will be removed soon.
}
\concept{deprecated}
