% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{get_rs_user_settings_names}
\alias{get_rs_user_settings_names}
\title{DEPRECATED. Use \code{rstudio_get_user_setting_names()} instead}
\usage{
get_rs_user_settings_names(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_get_user_setting_names]{rstudio_get_user_setting_names()}}}
}
\description{
\strong{DO NOT USE THIS FUNCTION!}
This function was replaced by \code{\link[=rstudio_get_user_setting_names]{rstudio_get_user_setting_names()}} and will be removed soon.
}
\concept{deprecated}
