% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_file_ids.R
\name{get_rstudio_file_ids}
\alias{get_rstudio_file_ids}
\alias{get_rstudio_file_ids_user}
\alias{get_rstudio_file_ids_project}
\title{Get RStudio document IDs and paths}
\usage{
get_rstudio_file_ids()

get_rstudio_file_ids_user()

get_rstudio_file_ids_project()
}
\value{
Data frame with columns \code{id} (for IDs) and \code{path} (for file paths).
}
\description{
Get RStudio file IDs and paths.
}
\examples{
\dontrun{\donttest{
head(get_rstudio_file_ids_user())
}}
}
\concept{paths and dirs}
