% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{get_rstudio_version_from_registry}
\alias{get_rstudio_version_from_registry}
\title{Get the RStudio version reported by the Windows installer registry entry}
\usage{
get_rstudio_version_from_registry()
}
\value{
A \code{\link[=numeric_version]{numeric_version()}} object, or \code{NULL} if not found.
}
\description{
Get the RStudio version reported by the Windows installer registry entry
}
