% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{get_vesion_bio}
\alias{get_vesion_bio}
\title{DEPRECATED. Use \code{get_vesion_pkg_bio()} instead}
\usage{
get_vesion_bio()
}
\description{
\strong{DO NOT USE THIS FUNCTION!}
This function was replaced by \code{\link[=get_vesion_pkg_bio]{get_vesion_pkg_bio()}} and will be removed
soon.
}
\concept{deprecated}
