% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git}
\alias{git}
\alias{get_git_core_editor_cmd}
\title{Helper functions for "Git".}
\usage{
get_git_core_editor_cmd(core_editor = "atom")
}
\arguments{
\item{core_editor}{(character) Name of Git core editor.}
}
\description{
Helper functions for "Git".
}
