% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjcfit.R
\name{hjcfit.fit}
\alias{hjcfit.fit}
\title{Something to do with qmatrix}
\usage{
hjcfit.fit(qmatrix)
}
\arguments{
\item{qmatrix}{your qmatrix.}
}
\value{
A fitted model.
}
\description{
Something to do with qmatrix
}
