% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjcfit.R
\name{hjcfit.fit_model}
\alias{hjcfit.fit_model}
\title{A maximum likelihood estimate for a Q-Matrix with a fixed number
of states. Computes the likelihood of a kinetic model given
observed bursts, with missed event correction, using the HJCFIT
library. This function is an interface for Log10Likelihood from
HJCFIT.}
\usage{
hjcfit.fit_model(qmatrix, nopen, bursts, resolution, nmax = 2,
  xtol = 10^(-10), rtol = 10^(-10), itermax = 100, REQMIN = 0.01,
  KONVGE = 10, KMAX = 10000)
}
\arguments{
\item{qmatrix}{your initial-guess qmatrix.}

\item{nopen}{number of open states.}

\item{bursts}{a list of bursts.}

\item{resolution}{Maximum length of the missed events / time resolution.}

\item{nmax}{The exact missed-event likelihood will be computed for 't < n_max * tau'}

\item{xtol}{Tolerance criteria for brentq().}

\item{rtol}{Tolerance criteria for brentq().}

\item{itermax}{Maximum number of iteration when calling brentq().}

\item{REQMIN}{Nelder-Mead parameter: stop if within required distance.}

\item{KONVGE}{Nelder-Mead parameter: check convergence every KONVGE steps.}

\item{KMAX}{Nelder-Mead parameter: maximum number of iterations.}
}
\value{
A qmatrix kinetic model which numerically maximizes likelihood.
}
\description{
https://github.com/DCPROGS/HJCFIT
}
\details{
See the link above and the references therein for details.
}
