% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjcfit.R
\name{hjcfit.likelihood}
\alias{hjcfit.likelihood}
\title{Computes the likelihood of a kinetic model given observed bursts,
with missed event correction, using the HJCFIT library. This
function is an interface for Log10Likelihood from HJCFIT.}
\usage{
hjcfit.likelihood(qmatrix, nopen, bursts, resolution, nmax = 2,
  xtol = 1e-10, rtol = 1e-10, itermax = 100)
}
\arguments{
\item{qmatrix}{your qmatrix.}

\item{nopen}{number of open states.}

\item{bursts}{a list of bursts.}

\item{resolution}{Maximum length of the missed events / time resolution.}

\item{nmax}{The exact missed-event likelihood will be computed for 't < n_max * tau'}

\item{xtol}{Tolerance criteria for brentq().}

\item{rtol}{Tolerance criteria for brentq().}

\item{itermax}{Maximum number of iteration when calling brentq().}
}
\value{
The -log10-likelihood of the model given the burst sequence.
}
\description{
https://github.com/DCPROGS/HJCFIT
}
\details{
See the link above and the references therein for details.
}
