% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{is_git_installed}
\alias{is_git_installed}
\title{Check if Git is installed}
\usage{
is_git_installed()
}
\value{
logical value.
}
\description{
Check if Git is installed
}
\examples{
is_git_installed()
}
\concept{programs}
