% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{is_rstudio_installed}
\alias{is_rstudio_installed}
\title{Check whether RStudio Desktop is installed, even if it is not running}
\usage{
is_rstudio_installed()
}
\value{
\code{TRUE}/\code{FALSE}.
}
\description{
Check whether RStudio Desktop is installed, even if it is not running
}
