% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages--find.R
\name{list_pkgs_used_in_files}
\alias{list_pkgs_used_in_files}
\title{List Packages Used in Specific File(s)}
\usage{
list_pkgs_used_in_files(files, exclude_base = TRUE, progress = FALSE, ...)
}
\arguments{
\item{files}{Character vector of file paths (R, Rmd, Qmd).}

\item{exclude_base}{Logical. If \code{TRUE} (default), base R packages
(e.g., \code{stats}, \code{utils}, \code{graphics}) are excluded from the output.}

\item{progress}{Logical. Whether to show a progress bar. Defaults to \code{FALSE}.}

\item{...}{Further arguments passed to \code{\link[renv:dependencies]{renv::dependencies()}}.}
}
\value{
Character vector of unique, sorted package names.
}
\description{
Scans specific R, Rmd, or Qmd file(s) and returns a vector of
unique package names used inside them.
}
\concept{packages}
