% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{meld}
\alias{meld}
\alias{browse_meld_homepage}
\alias{download_meld}
\alias{get_default_path_to_meld}
\alias{is_meld_installed}
\alias{set_meld_as_git_difftool_mergetool}
\title{Helper functions for "Meld"}
\usage{
browse_meld_homepage()

download_meld()

get_default_path_to_meld()

is_meld_installed(path_to_meld = get_default_path_to_meld())

set_meld_as_git_difftool_mergetool(
  path_to_meld = NULL,
  prompt = FALSE,
  keepBackup = FALSE,
  trustExitCode = TRUE
)
}
\arguments{
\item{path_to_meld}{(character|\code{NULL})}

\item{prompt}{(logical)}

\item{keepBackup}{(logical)}

\item{trustExitCode}{(logical)}
}
\description{
Helper functions for "Meld"
}
\examples{
is_meld_installed()
}
\concept{programs}
