% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{open_backup_dir}
\alias{open_backup_dir}
\title{Open directory with backup files}
\usage{
open_backup_dir(...)
}
\arguments{
\item{...}{Subdirectories name.}
}
\description{
Open directory with backup files
}
