% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-functions.R
\name{open_in_rs}
\alias{open_in_rs}
\title{DEFUNCT. Use \code{open_in_rstudio()} instead}
\usage{
open_in_rs(path)
}
\arguments{
\item{path}{See \code{\link[=open_in_rstudio]{open_in_rstudio()}}.}
}
\description{
\strong{DO NOT USE THIS FUNCTION!}
This function was replaced by \code{\link[=open_in_rstudio]{open_in_rstudio()}} and will be removed soon.
}
\concept{defunct}
