% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{open_r_environ}
\alias{open_r_environ}
\alias{get_path_r_environ}
\title{Open \code{.Renviron} File}
\usage{
get_path_r_environ(scope = c("user", "project"))

open_r_environ()
}
\arguments{
\item{scope}{(character) The scope of file. One of "user" or "project".}
}
\description{
Functions to get path to and open \code{.Renviron} file that contains
definitions of R environment variables.

Compared to  \code{\link[usethis:edit_r_environ]{usethis::edit_r_environ()}}, \code{open_r_environ()} does not create
file if it does not exist.
}
\examples{
get_path_r_environ()

}
\seealso{
\itemize{
\item \code{\link[usethis:edit_r_environ]{usethis::edit_r_environ()}}
}
}
\concept{paths and dirs}
