% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_order_to_install.R
\name{optimize_order_to_install}
\alias{optimize_order_to_install}
\title{optimize_order_to_install}
\usage{
optimize_order_to_install(
  pkgs_vec = get_pkgs_recommended()$paketas,
  recursive_dependencies = TRUE
)
}
\arguments{
\item{pkgs_vec}{???}

\item{recursive_dependencies}{(logical) ???}
}
\value{
???
}
\description{
Helpeer function that suggest how to optimize order of packages in the
vector of packages in order not to repeat installation of
the same packages.
}
\details{
recursive_dependencies = TRUE requires internet connection.
}
