% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander_options.R
\name{pander_options}
\alias{pander_options}
\title{Get or set Pander options.}
\usage{
pander_options(o, value, ...)
}
\arguments{
\item{o}{(string) option name.}

\item{value}{value to assign (optional)}

\item{...}{valid option - value pairs (e.g., \code{option = "value"}).}
}
\description{
Get or set Pander options.
}
\examples{
p_opts <- pander_options()
head(p_opts, n = 2)

pander_options("digits")

pander_options("digits", 5)
pander_options("digits")

pander_options(digits = 2)
pander_options("digits")

}
\seealso{
\code{\link[pander:evalsOptions]{pander::evalsOptions()}}
}
