% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{parse_proj_path}
\alias{parse_proj_path}
\alias{extract_proj_name}
\title{Parse project's path}
\usage{
parse_proj_path(proj_path)

extract_proj_name(proj_path)
}
\arguments{
\item{proj_path}{(character) Path to \verb{*.Rproj} file.}
}
\value{
\itemize{
\item \code{parse_proj_path()} returns data frame with columns \code{name} (with project
name, i.e., folder name) and \code{path} to *.Rproj file.
\item \code{extract_proj_name()} (character) a vector of extracted character names.
}
}
\description{
Get details about RStudio project from its path.
\itemize{
\item \code{extract_proj_name()} extracts RStudio project's name form a path to the
project.
}
}
\examples{
path_to_project <-  c("list/proj.Rproj", "C:/R/BS-2020/bs.Rproj",
    "C:/data/analysis/proj.Rproj")

extract_proj_name(path_to_project)

parse_proj_path(path_to_project)
}
\concept{rstudio projects}
