% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings--preferences.R
\name{print_pref_diff_summary}
\alias{print_pref_diff_summary}
\title{Print a concise summary of \code{summarize_pref_diff()} output.}
\usage{
print_pref_diff_summary(diff_df, x_arg, y_arg)
}
\description{
Print a concise summary of \code{summarize_pref_diff()} output.
}
\keyword{internal}
