% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{qmatrix.create}
\alias{qmatrix.create}
\title{Create a Q-Matrix object from a kinetic rate matrix and vector
of associated conductance states.}
\usage{
qmatrix.create(mat, states)
}
\arguments{
\item{mat}{The kinetic rate matrix.}

\item{states}{The associated conductance states.}
}
\value{
the Q-Matrix object (ready for HJCFIT)
}
\description{
Note that the order of the states will be re-arranged, so that the
open states are the in the top-left of the matrix.
}
\examples{

m = matrix(c(
 -3050,  3050,   0,       0,       0,       0,      0,      0,
 910,    -4755,  3500,    345,     0,       0,      0,      0,
 0,      15000,  -23400,  0,       08400,   0,      0,      0,
 0,      24100,  0,       -87800,  54500,   9200,   0,      0,
 0,      0,      86400,   17200,   -28600,  0,      25000,  0,
 0,      0,      0,       7300,    0,       -7300,  0,      0,
 0,      0,      0,       0,       100,     0,      -4600,  4500,
 0,      0,      0,       0,       0,       0,      3000,   -3000
), ncol=8, byrow=TRUE)

qm <- qmatrix.create(m, c(0,0,0,0,0,1,1,0))

qm
}
