% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{qmatrix.num_open}
\alias{qmatrix.num_open}
\title{Return the number of open states.}
\usage{
qmatrix.num_open(mat)
}
\arguments{
\item{mat}{The Q-matrix.}
}
\value{
the number of open states
}
\description{
Return the number of open states.
}
\examples{

# import some of the data included with the package
infile <- system.file("extdata", "example5.bin", package = "scbursts")

qm <- qmatrix.read(infile)
qm

qmatrix.states(qm)
qmatrix.num_open(qm)
}
