% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{qmatrix.num_states}
\alias{qmatrix.num_states}
\title{Return the number of states (size of the matrix)}
\usage{
qmatrix.num_states(mat)
}
\arguments{
\item{mat}{The Q-matrix.}
}
\value{
the number of states
}
\description{
Return the number of states (size of the matrix)
}
\examples{

# import some of the data included with the package
infile <- system.file("extdata", "example5.bin", package = "scbursts")

qm <- qmatrix.read(infile)
qm

qmatrix.states(qm)
qmatrix.num_states(qm)
}
