% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{qmatrix.read}
\alias{qmatrix.read}
\title{Read a Q-Matrix from a .bin file}
\usage{
qmatrix.read(filename)
}
\arguments{
\item{filename}{The filename}
}
\value{
The Q-Matrix
}
\description{
Reads one of the .bin model files from Mat's simulator
}
\examples{

#' # import some of the data included with the package
infile <- system.file("extdata", "example5.bin", package = "scbursts")

qm <- qmatrix.read(infile)
qm
}
