% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{qmatrix.reorder}
\alias{qmatrix.reorder}
\title{Reorder the states of a matrix so that
the top left corner contains all the open states.}
\usage{
qmatrix.reorder(mat)
}
\arguments{
\item{mat}{The matrix.}
}
\value{
The Q-Matrix
}
\description{
Necessary for HJCFIT.
}
\examples{

# import some of the data included with the package
infile <- system.file("extdata", "example5.bin", package = "scbursts")

qm <- qmatrix.read(infile)
qm

qm_r <- qmatrix.reorder(qm)
qm_r
}
