% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{qmatrix.sanity_check}
\alias{qmatrix.sanity_check}
\title{Verify that this is a valid Q-Matrix.}
\usage{
qmatrix.sanity_check(mat, no_order = FALSE)
}
\arguments{
\item{mat}{The Q-matrix.}
}
\value{
TRUE if the matrix is ok, FALSE and a warning message otherwise.
}
\description{
Verify that this is a valid Q-Matrix.
}
\examples{

# import some of the data included with the package
infile <- system.file("extdata", "example5.bin", package = "scbursts")

qm <- qmatrix.read(infile)
qm

qmatrix.sanity_check(qm)

# entry (3,3) is positive! It should be negative.
m = matrix(c(
 -3050,  3050,   0,       0,       0,       0,      0,      0,
 910,    -4755,  3500,    345,     0,       0,      0,      0,
 0,      15000,  23400,   0,       08400,   0,      0,      0,
 0,      24100,  0,       -87800,  54500,   9200,   0,      0,
 0,      0,      86400,   17200,   -28600,  0,      25000,  0,
 0,      0,      0,       7300,    0,       -7300,  0,      0,
 0,      0,      0,       0,       100,     0,      -4600,  4500,
 0,      0,      0,       0,       0,       0,      3000,   -3000
), ncol=8, byrow=TRUE)

# qmatrix.create calls sanity check for you, too.
qm <- qmatrix.create(m, c(0,0,0,0,0,1,1,0))
qm

qmatrix.sanity_check(qm)
}
