% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{qmatrix.states}
\alias{qmatrix.states}
\title{Return the vector of states}
\usage{
qmatrix.states(mat)
}
\arguments{
\item{mat}{The Q-matrix.}
}
\value{
the states vector
}
\description{
Return the vector of states
}
\examples{

# import some of the data included with the package
infile <- system.file("extdata", "example5.bin", package = "scbursts")

qm <- qmatrix.read(infile)
qm

qmatrix.states(qm)
}
