% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{R-related-dirs}
\alias{R-related-dirs}
\alias{get_path_r_user_dir}
\alias{open_r_user_dir}
\title{Directories of R-Related Files}
\usage{
get_path_r_user_dir(...)

open_r_user_dir()
}
\description{
Directories of R-Related files.
\itemize{
\item \code{get_path_r_user_dir()}-- gets path to the main R user directory.
}
}
\examples{
get_path_r_user_dir()

}
\seealso{
\itemize{
\item \code{\link[fs:file_show]{fs::file_show()}}, \code{\link[=browseURL]{browseURL()}},
\item \code{\link[rstudioapi:navigateToFile]{rstudioapi::navigateToFile()}},
\item \code{\link[utils:file.edit]{utils::file.edit()}}
}
}
\concept{paths and dirs}
