% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED-functions.R
\name{read_rs_ui_prefs}
\alias{read_rs_ui_prefs}
\title{DEPRECATED. Use \code{rstudio_read_ui_prefs()} instead}
\usage{
read_rs_ui_prefs(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_read_ui_prefs]{rstudio_read_ui_prefs()}}}
}
\description{
DO NOT USE THIS FUNCTION!
This function was replaced by \code{\link[=rstudio_read_ui_prefs]{rstudio_read_ui_prefs()}} and will be removed soon.
}
\concept{deprecated}
