% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{read_rs_user_settings}
\alias{read_rs_user_settings}
\title{DEPRECATED. Use \code{rstudio_read_user_settings()} instead}
\usage{
read_rs_user_settings(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_read_user_settings]{rstudio_read_user_settings()}}}
}
\description{
DO NOT USE THIS FUNCTION!
This function was replaced by \code{\link[=rstudio_read_user_settings]{rstudio_read_user_settings()}} and will be removed soon.
}
\concept{deprecated}
