% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{reload_rstudio}
\alias{reload_rstudio}
\title{DEFUNCT. Use \code{rstudio_reload_ui()} instead}
\usage{
reload_rstudio()
}
\description{
\strong{DO NOT USE THIS FUNCTION!}
This function was replaced by \code{\link[=rstudio_reload_ui]{rstudio_reload_ui()}} and will be removed soon.
}
\concept{defunct}
