% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{reset_rstudio_gmc}
\alias{reset_rstudio_gmc}
\title{DEPRECATED. Use \code{rstudio_gmc_reset()} instead}
\usage{
reset_rstudio_gmc(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_gmc_reset]{rstudio_gmc_reset()}}}
}
\description{
This function was replaced by \code{\link[=rstudio_gmc_reset]{rstudio_gmc_reset()}} and will be removed soon.
}
\concept{deprecated}
