% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{reset_rstudio_layout}
\alias{reset_rstudio_layout}
\title{DEPRECATED. Use \code{rstudio_reset_layout()} instead}
\usage{
reset_rstudio_layout(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_reset_layout]{rstudio_reset_layout()}}}
}
\description{
DO NOT USE THIS FUNCTION!
This function was replaced by \code{\link[=rstudio_reset_layout]{rstudio_reset_layout()}} and will be removed soon.
}
\concept{deprecated}
