% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-functions.R
\name{reset_rstudio_user_settings}
\alias{reset_rstudio_user_settings}
\title{DEFUNCT. Use \code{rstudio_reset_user_settings()} instead}
\usage{
reset_rstudio_user_settings(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_reset_user_settings]{rstudio_reset_user_settings()}}}
}
\description{
\strong{DO NOT USE THIS FUNCTION!}
This function was replaced by \code{\link[=rstudio_reset_user_settings]{rstudio_reset_user_settings()}} and will be removed soon.
}
\concept{defunct}
