% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED-functions.R
\name{restart_r}
\alias{restart_r}
\title{DEPRECATED. Use \code{rstudio_restart_r()} instead}
\usage{
restart_r(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_restart_r]{rstudio_restart_r()}}}
}
\description{
This function was replaced by \code{\link[=rstudio_restart_r]{rstudio_restart_r()}} and will be removed soon.
}
\concept{deprecated}
