% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{rproj_open_project}
\alias{rproj_open_project}
\title{Open RStudio project.}
\usage{
rproj_open_project(
  name = NULL,
  new_session = FALSE,
  proj_list = NULL,
  proj_list_path = rproj_get_path_rs_recent_proj_list(),
  only_existing = TRUE
)
}
\arguments{
\item{name}{(string|\code{NULL}) The name of the project or \code{NULL} to choose
a project interactively.}

\item{new_session}{(logical) should the project be opened in a new session,
or should the current RStudio session switch to that project? Note
that \code{TRUE} values are only supported with RStudio Desktop and RStudio
Server Pro.}

\item{proj_list}{(data frame) The result of \code{\link[=rproj_read_projects]{rproj_read_projects()}} or \code{NULL}.}

\item{proj_list_path}{(string) The path to the file with the list of project
paths. If \code{proj_list} is not \code{NULL}, then \code{proj_list_path} is ignored.}

\item{only_existing}{(logical) If \code{TRUE}, non-existing projects and projects
with broken paths are removed from the list of choices.}
}
\value{
Opens the indicated project.
}
\description{
Open RStudio project by name or interactively.
}
\examples{
\donttest{\dontrun{

rproj_open_project()

rproj_open_project("bs")

rproj_open_project("R-2019-project")

}}
}
\seealso{
\itemize{
\item \code{\link[rstudioapi:openProject]{rstudioapi::openProject()}}
\item \code{\link[rstudioapi:initializeProject]{rstudioapi::initializeProject()}}
}
}
