% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{RStudio-config-dir}
\alias{RStudio-config-dir}
\alias{get_path_r_user_dir}
\alias{get_path_rs_config_dir}
\alias{get_path_rs_desktop_config_dir}
\alias{get_path_rs_snippets_dir}
\alias{get_path_rs_keybindings_dir}
\title{Directories of R and RStudio (desktop) settings, preferences, etc.}
\usage{
get_path_r_user_dir(...)

get_path_rs_config_dir(...)

get_path_rs_desktop_config_dir(..., .check = FALSE)

get_path_rs_snippets_dir()

get_path_rs_keybindings_dir()
}
\arguments{
\item{...}{(character) Parts of the path. Path to sub-directories.}

\item{.check}{(logical) If \code{TRUE}, additionally checks for path existance.}
}
\value{
\itemize{
\item \code{get_path_rs_config_dir()} (string) path, if it exists, or error.
}

(string) path to RStudio desktop user settings directory.
When \code{.check = TRUE}, renturns error, if the path does not exist.
}
\description{
\itemize{
\item \code{get_path_r_user_dir()}-- gets path to the main R user directory.
}

\itemize{
\item \code{get_path_rs_config_dir()} -- gets path to RStudio configuration directory
or its sub-directories.
}

\itemize{
\item `get_path_rs_desktop_config_dir()`` - gets path to directory (and its
sub-directories) of RStudio desktop user interface settings.
}
}
\examples{
get_path_r_user_dir()

\dontrun{\donttest{
get_path_rs_config_dir()
}}
\dontrun{\donttest{
get_path_rs_desktop_config_dir()

get_path_rs_desktop_config_dir("dictionaries")
}}
get_path_rs_snippets_dir()

get_path_rs_keybindings_dir()

}
\seealso{
\itemize{
\item \code{get_path_rs_desktop_config_dir()}:
https://support.rstudio.com/hc/en-us/articles/200534577-Resetting-RStudio-Desktop-s-State
}
}
\concept{paths and dirs}
