% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{rstudio_activate_console}
\alias{rstudio_activate_console}
\title{Activate the console in RStudio when available.}
\usage{
rstudio_activate_console()
}
\value{
Invisibly returns \code{NULL} when RStudio is unavailable.
}
\description{
Activate the console in RStudio when available.
}
\examples{
if (interactive()) {
  rstudio_activate_console()
}
}
\keyword{internal}
