% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{rstudio_clear_console_ask}
\alias{rstudio_clear_console_ask}
\title{Ask before clearing the RStudio console.}
\usage{
rstudio_clear_console_ask()
}
\value{
Invisibly returns \code{NULL} if RStudio is unavailable or the user says no.
}
\description{
Ask before clearing the RStudio console.
}
\examples{
\dontrun{
rstudio_clear_console_ask()
}
}
\keyword{internal}
