% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment.subconductance_as}
\alias{segment.subconductance_as}
\title{Imposes a fixed conductance level (0 or 1) to all dwells with subconductance levels.}
\usage{
segment.subconductance_as(segment, level)
}
\arguments{
\item{segment, }{the segment containing dwells and states.}

\item{level, }{either 'open' or 'closed'}
}
\value{
A modified copy of the original segment
}
\description{
The user specifies the desired level ('open' or 'closed'). The function will modify
any subconductance level (that is not 0 or 1) to be the desired level 1 for 'open'
or 0 for 'closed'. The function then reutrns a collapsed segment containing 
alternating dwells.
(See segment.consecutives_to_dwells for details about the collapsed segment.)
}
