% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{set_rstudio_keybindings}
\alias{set_rstudio_keybindings}
\title{Set RStudio keybindings}
\usage{
set_rstudio_keybindings(which = "none", backup = TRUE)
}
\arguments{
\item{which}{(string) The type of keybindings. Currently available value is
\code{"bio-default"}.}

\item{backup}{(logical) If \code{TRUE}, a back-up copy of files with current
keybindings is ctreated.}
}
\description{
Set RStudio keybindings
}
\examples{
\dontrun{\dontest{

bio::set_rstudio_keybindings(which = "bio-default")
bio::restart_rstudio()

}}
}
