% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{set_wd}
\alias{set_wd}
\alias{set_wd_interactive}
\alias{set_wd_to_active_doc}
\alias{set_wd_to_files_pane}
\alias{set_wd_to_project_dir}
\title{Change working directory in RStudio}
\usage{
set_wd_interactive()

set_wd_to_active_doc()

set_wd_to_files_pane()

set_wd_to_project_dir()
}
\description{
Functions do the same as RStudio menus
RStudio menu -> Session -> Set Working Directory -> ...:
}
\details{
\itemize{
\item \code{set_wd_to_active_doc()}: ... -> To Source File Location
\item \code{set_wd_to_files_pane()}: ... -> to Files Pane Directory
(directory opened in files tab)
\item \code{set_wd_to_project_dir()}: ... -> To Project Directory
\item \code{set_wd_interactive()}: ... -> Choose Directory...
}
}
\concept{r and rstudio settings}
