% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_move}
\alias{str_move}
\alias{str_move_before}
\alias{str_move_after}
\title{Change the position of a string in a vector}
\usage{
str_move_before(x, what, move_before)

str_move_after(x, what, move_after)
}
\arguments{
\item{x}{(character vector)}

\item{what}{(string)}

\item{move_before}{(string)}

\item{move_after}{(sting)}
}
\description{
Change the position of a string in a character vector.
}
\examples{
x <- c("a", "b", "d", "e", "c", "f")
str_move_before(x, what = "c", move_before = "d")

x <- c("f", "c", "e", "d", "b", "a")
str_move_after(x, what = "c", move_after = "d")

x <- c("a", "b", "d", "e", "c", "f")
str_move_before(x, what = "c", move_before = "d")

x <- c("f", "c", "e", "d", "b", "a")
str_move_after(x, what = "c", move_after = "d")
}
