% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{suggest_optimized_order_of_packages}
\alias{suggest_optimized_order_of_packages}
\title{Optimize order of packages to install.}
\usage{
suggest_optimized_order_of_packages(pkgs_vec, recursive_dependencies = FALSE)
}
\arguments{
\item{pkgs_vec}{(character) Vector of package names.}

\item{recursive_dependencies}{(logical) If \code{TRUE}, recursive dependencies are
also checked.
\strong{NOTE:} \code{recursive_dependencies = TRUE}requires internet connection.}
}
\value{
A list with suggestion, which packages should be moved.
}
\description{
Helper function that takes a vector of package names and suggests how to
sort the packages in order not to repeat installation of the same
package if it can be installed as a dependence of some other package.
}
\examples{
\dontrun{\donttest{
suggest_optimized_order_of_packages(c("stringr", "stringi", "glue", "readr"))
}}
}
