% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_task_names.R
\name{to_ascii_lower}
\alias{to_ascii_lower}
\title{Transliterate to ASCII Lower-case Symbol}
\usage{
to_ascii_lower(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Transliterate to ASCII Lower-case Symbol
}
\examples{
to_ascii_lower("AbD")
}
