% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_task_names.R
\name{u_prepare_assignment_dc}
\alias{u_prepare_assignment_dc}
\title{Prepare Files For Submission (DataCamp Task)}
\usage{
u_prepare_assignment_dc(
  uzduoties_nr,
  varianto_nr,
  specialybe,
  pavarde,
  vardas,
  sertifikatas,
  konspektas,
  open_dir = TRUE,
  emokymai_id = NULL
)
}
\arguments{
\item{uzduoties_nr}{(integer)}

\item{varianto_nr}{(integer)}

\item{specialybe}{(string)}

\item{pavarde}{(string)}

\item{vardas}{(string)}

\item{sertifikatas}{(string) Path to PDF file
(certificate issued by DataCamp).}

\item{konspektas}{(string) Path to PDF file.}

\item{open_dir}{(\strong{\code{TRUE}}|\code{FALSE}) Open the directory with submission files.}

\item{emokymai_id}{(integer) Open the directory with submission files.}
}
\description{
The function creates a folder (e.g.,\code{pateikti-uzduoti-U00}) and copies files
indicated in \code{sertifikatas} and \code{konspektas} to that folder.
The copied files are properly renamed.
If needed, the function opens the folder and assignment submission webpage.
}
\examples{
\donttest{\dontrun{
u_prepare_assignment_dc(
  uzduoties_nr = 1,
  varianto_nr  = 000,
  specialybe   = "molekuline biologija",
  pavarde      = "Pavarde",
  vardas       = "Vardas Antras",
  sertifikatas = "pratybos/certificate.pdf",
  konspektas   = "pratybos/konspektas.pdf"
)
}}
}
