% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_tasks.R
\name{u_task}
\alias{u_task}
\alias{u_create_filename}
\alias{u_parse_filename}
\alias{u_check_filename}
\title{Create Filename for Type U Task}
\usage{
u_create_filename(
  uzduoties_nr,
  varianto_nr = NULL,
  specialybe = NULL,
  pavarde = NULL,
  vardas = NULL,
  patikslinimas = "",
  dokumento_formatas = NULL
)

u_parse_filename(x)

u_check_filename(x)
}
\arguments{
\item{uzduoties_nr}{(integer) The number of U task.}

\item{varianto_nr}{(integer) The variant of the task.
Each student has a unique personal variant number.}

\item{specialybe}{(character) Name of study programme.}

\item{pavarde}{(character) Family name (names).}

\item{vardas}{(character) Sure name (names).}

\item{patikslinimas}{(character) Additional information.}

\item{dokumento_formatas}{(character) File name extension.}

\item{x}{(character) A file name to parse or to check if it's structure
is correct.}
}
\value{
\itemize{
\item \code{u_create_filename()} returns sting with formatted file name.
\item \code{u_parse_filename()} returns data frame with details extracted from
the filename.
\item \code{u_check_filename()} returns \code{TRUE} if the structure of file name is
correct and \code{FALSE} otherwise.
}
}
\description{
Create Filename for Type U Task
}
\examples{
u_create_filename(
  uzduoties_nr       = 1,
  varianto_nr        = 000,
  specialybe         = "molekuline biologija",
  pavarde            = "Pavarde",
  vardas             = "Vardas Antras",
  patikslinimas      = "sertifikatas",
  dokumento_formatas = "pdf"
)

u_check_filename("U01-v000-[molbio]-[Pavarde]-[Vardas-Antras]-sertifikatas.pdf")
u_parse_filename("U01-v000-[molbio]-[Pavarde]-[Vardas-Antras]-sertifikatas.pdf")

u_parse_filename("U03-v000-[molekuline]-[Pavarde]-[Vardas].zip")

# Incorrect family name:
u_parse_filename("U03-v000-[molekuline]-[Pavarde3]-[Vardas].zip")
}
