% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_update.R
\name{update_pkg_rcmdr_biostat}
\alias{update_pkg_rcmdr_biostat}
\alias{check_updates_rcmdr_biostat}
\title{Update '\pkg{RcmdrPlugin.biostat}' and related packages}
\usage{
check_updates_rcmdr_biostat(
  show_status = "always",
  install = "outdated",
  upgrade = TRUE,
  ...
)

update_pkg_rcmdr_biostat(upgrade = TRUE, force = FALSE, ...)
}
\description{
Update '\pkg{RcmdrPlugin.biostat}' and related packages
}
\examples{
if (interactive()) {
  update_pkg_rcmdr_biostat()
  check_updates_rcmdr_biostat()
}
}
\concept{packages}
