% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_update.R
\name{update_pkg_snippets}
\alias{update_pkg_snippets}
\title{Update '\pkg{snippets}' and related packages}
\usage{
update_pkg_snippets(upgrade = TRUE, force = FALSE, ...)
}
\description{
Update '\pkg{snippets}' and related packages
}
\examples{
if (interactive()) {
  update_pkg_snippets()
}
}
\concept{packages}
