% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio-related.R
\name{update_rcmdr_biostat}
\alias{update_rcmdr_biostat}
\title{Update Rcmdr Biostat}
\usage{
update_rcmdr_biostat()
}
\description{
Update Rcmdr Biostat
}
\examples{
\dontrun{\donttest{

update_rcmdr_biostat()

}}
}
