% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{version-control}
\alias{version-control}
\alias{vc_open_rs_proj_vc_opts}
\alias{vc_show_rs_rsa_key}
\alias{vc_browse_verson_control_help}
\title{Version control in RStudio.}
\usage{
vc_open_rs_proj_vc_opts()

vc_show_rs_rsa_key()

vc_browse_verson_control_help()
}
\description{
Helper functions for version control in RStudio.
}
