% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{atom}
\alias{atom}
\alias{browse_atom_homepage}
\alias{is_atom_installed}
\alias{set_atom_as_git_core_editor}
\alias{install_atom_meld}
\title{Helper functions for "Atom".}
\usage{
browse_atom_homepage()

is_atom_installed()

set_atom_as_git_core_editor()

install_atom_meld()
}
\description{
Helper functions for "Atom".
}
\examples{
is_atom_installed()
}
\concept{programs}
