% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio-package.R
\docType{package}
\name{bio-package}
\alias{bio}
\alias{bio-package}
\title{bio: Resources Manager for BS-2021 and R-2021}
\description{
Resources manager for course units BS-2020 and R-2020.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mokymai.github.io/bio/}
  \item \url{https://github.com/mokymai/bio/}
  \item Report bugs at \url{https://github.com/mokymai/bio/issues}
}

}
\author{
\strong{Maintainer}: Vilmantas Gegzna \email{GegznaV@gmail.com} (\href{https://orcid.org/0000-0002-9500-5167}{ORCID})

}
\keyword{internal}
