% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio-related.R
\name{bio_version}
\alias{bio_version}
\alias{get_vesion_pkg_bio}
\alias{check_updates_pkg_bio}
\alias{update_pkg_bio}
\title{Version management of package '\pkg{bio}'}
\usage{
get_vesion_pkg_bio()

check_updates_pkg_bio(
  show_status = "always",
  install = "outdated",
  upgrade = FALSE,
  ...
)

update_pkg_bio(upgrade = TRUE, force = FALSE, quiet = TRUE)
}
\arguments{
\item{show_status}{See argument \code{show_status} in \code{\link[=get_pkgs_installation_status]{get_pkgs_installation_status()}}.}

\item{install}{See argument \code{install} in \code{\link[=get_pkgs_installation_status]{get_pkgs_installation_status()}}.}

\item{upgrade}{Upgrade dependencies.
See \code{upgrade} in \code{\link[remotes:install_cran]{remotes::install_cran()}}.}

\item{...}{Arguments to further methods.}

\item{force}{(logical) \code{TRUE} or \code{FALSE}.
Force to update.
See \code{force} in \code{\link[remotes:install_cran]{remotes::install_cran()}}.}

\item{quiet}{(logical) \code{TRUE} or \code{FALSE}.
Should installation messages be printed?
See \code{quiet} in \code{\link[remotes:install_cran]{remotes::install_cran()}}.}
}
\description{
\code{get_vesion_pkg_bio()} returns version number of package \pkg{bio}.
\code{check_updates_pkg_bio()} checks if package \pkg{bio} has a recommended version.
\code{update_pkg_bio()} tries to update package \pkg{bio}.
}
\examples{
get_vesion_pkg_bio()

\dontrun{\donttest{

check_updates_pkg_bio()
update_bio()

}}
}
\concept{packages}
