% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{check_installed_programs}
\alias{check_installed_programs}
\alias{check_user_info}
\title{Check the versions of required programs}
\usage{
check_installed_programs(which = "main", skip_online_check = FALSE)

check_user_info()
}
\arguments{
\item{which}{(character) Which programs should be checked? Options:
\code{main}, \code{all}, \code{BS-2020}.}

\item{skip_online_check}{(logical) If \code{TRUE}, the numbers of newest available
stable programs are downloaded, when internet connection is connected.}
}
\value{
The results of version checking is printed.
}
\description{
Check the versions of required programs
}
\examples{

\dontrun{\donttest{

check_user_info()

check_installed_programs()

check_installed_programs("all")

check_installed_programs("BS-2020")

}}
}
\concept{programs}
