% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{check_installed_programs}
\alias{check_installed_programs}
\alias{check_user_info}
\title{Check required programs and user-related information}
\usage{
check_installed_programs(type = "main", skip_online_check = FALSE)

check_user_info()
}
\arguments{
\item{type}{(character) Which programs should be checked? Options:
\code{main}, \code{all}, \code{gmc-bs}, \code{gmc-r}.}

\item{skip_online_check}{(logical) If \code{TRUE}, the numbers of newest available
stable programs are downloaded, when internet connection is connected.}
}
\value{
Invisible \code{NULL}.
The results of probram checking are printed.
}
\description{
Check absence/presence and (in some cases) versions of required programs
and user-related information.
}
\examples{

\dontrun{\donttest{

check_user_info()

check_installed_programs()

check_installed_programs("all")

}}
}
\concept{programs}
