% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{check_updates_bio}
\alias{check_updates_bio}
\title{DEPRECATED. Use \code{check_updates_pkg_bio()} instead}
\usage{
check_updates_bio(
  show_status = "always",
  install = "outdated",
  upgrade = FALSE,
  ...
)
}
\arguments{
\item{show_status}{See \code{\link[=check_updates_pkg_bio]{check_updates_pkg_bio()}}.}

\item{install}{See \code{\link[=check_updates_pkg_bio]{check_updates_pkg_bio()}}.}

\item{upgrade}{See \code{\link[=check_updates_pkg_bio]{check_updates_pkg_bio()}}.}

\item{...}{See \code{\link[=check_updates_pkg_bio]{check_updates_pkg_bio()}}.}
}
\description{
\strong{DO NOT USE THIS FUNCTION!}
This function was replaced by \code{\link[=check_updates_pkg_bio]{check_updates_pkg_bio()}} and will be removed
soon.
}
\concept{deprecated}
