% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programs.R
\name{check_user_info}
\alias{check_user_info}
\title{Get User-Related Information}
\usage{
check_user_info()
}
\description{
Get user-related information.
}
\examples{
\dontrun{\donttest{
check_user_info()
}}
}
\concept{check}
\concept{check-user-info}
