% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{clear_and_reset}
\alias{clear_and_reset}
\alias{clear_r_history}
\alias{clear_rs_history}
\alias{clear_r_workspace}
\alias{reset_rstudio_layout}
\title{Clear and Reset R and RStudio.}
\usage{
clear_r_history(backup = TRUE)

clear_rs_history()

clear_r_workspace()

reset_rstudio_layout(rs_layout = "left")
}
\arguments{
\item{backup}{(logical) If \code{TRUE}, a backup copy is created.}

\item{rs_layout}{(\code{"left"}|\code{"right"}) Type of RStudio panes layout.}
}
\description{
Clear and Reset R and RStudio settings and preferences.
}
