% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{compare_version}
\alias{compare_version}
\title{Compace version numbers.}
\usage{
compare_version(v_installed, v_required)
}
\arguments{
\item{v_installed}{vector with version numbers}

\item{v_required}{vector with version numbers}
}
\value{
The same as in \code{\link[utils:compareVersion]{utils::compareVersion()}}, just a vector.
}
\description{
Compace version numbers.
}
\examples{
compare_version("2.4", "2")
compare_version("2.3", "2.3")
compare_version("2.3", "2.3.1")
}
