% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_instalation_status}
\alias{get_pkgs_instalation_status}
\title{Get infromation about recommended packages}
\usage{
get_pkgs_instalation_status()
}
\value{
Data frame

Data frame with ???
}
\description{
Get infromation about recommended packages
}
\examples{
head(get_pkgs_instalation_status())
}
