% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_instalation_status}
\alias{get_pkgs_instalation_status}
\title{get_pkgs_instalation_status}
\usage{
get_pkgs_instalation_status(
  pkgs_rec = get_pkgs_recommended(),
  pkgs_inst = get_pkgs_installed(),
  pkgs_req_v = get_pkgs_req_version()
)
}
\value{
Data frame with ???
}
\description{
get_pkgs_instalation_status
}
\examples{
\dontrun{\donttest{

head(get_pkgs_instalation_status())

}}
}
