% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_installation_status_local}
\alias{get_pkgs_installation_status_local}
\alias{get_pkgs_installation_status}
\alias{print.pkgs_installation_status}
\alias{get_pkgs_installation_code}
\alias{get_pkgs_installation_code_cran}
\alias{get_pkgs_installation_code_github}
\title{Get package installation status.}
\usage{
get_pkgs_installation_status_local(which, local_list = TRUE)

get_pkgs_installation_status(which, show = "required", local_list = FALSE)

\method{print}{pkgs_installation_status}(x, ...)

get_pkgs_installation_code(x)

get_pkgs_installation_code_cran(x)

get_pkgs_installation_code_github(x)
}
\arguments{
\item{show}{(character)
\itemize{
\item If \code{"required"}, only packages, that need to be installed/updated are shown.
}}

\item{x}{Object of interest.}

\item{...}{Arguments to other methods.}
}
\description{
Get package installation status.
}
\examples{
\dontrun{\donttest{

head(get_pkgs_installation_status_local("r209"))

}}
\dontrun{\donttest{

which <- "r209"

get_pkgs_installation_status("r209")
(status_all <- get_pkgs_installation_status("r209", show = "all"))

get_pkgs_installation_code(status_all)
get_pkgs_installation_code_cran(status_all)
get_pkgs_installation_code_github(status_all)
}}
}
\seealso{
Other R-packages-related functions: 
\code{\link{compare_version}()},
\code{\link{get_pkgs_cran_details}()},
\code{\link{get_pkgs_installed}()},
\code{\link{get_pkgs_non_cran_installation_details}()},
\code{\link{get_pkgs_recommended}()},
\code{\link{get_pkgs_req_version}()}
}
\concept{R-packages-related functions}
