% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_installation_status_local}
\alias{get_pkgs_installation_status_local}
\alias{get_pkgs_installation_status}
\alias{print.pkgs_installation_status}
\alias{get_pkgs_installation_code}
\alias{get_pkgs_installation_code.pkgs_installation_status}
\alias{get_pkgs_installation_code_cran}
\alias{get_pkgs_installation_code_github}
\alias{get_pkgs_installation_code_other}
\title{Get package installation status and code.}
\usage{
get_pkgs_installation_status_local(
  list_name,
  use_local_list = getOption("bio.use_local_list", TRUE)
)

get_pkgs_installation_status(
  list_name,
  include = "outdated",
  show_status = include,
  install = include,
  cran = install,
  github = install,
  elsewhere = install,
  use_local_list = getOption("bio.use_local_list", FALSE)
)

\method{print}{pkgs_installation_status}(x, ...)

get_pkgs_installation_code(x, ...)

\method{get_pkgs_installation_code}{pkgs_installation_status}(x, ..., to_clipboard = FALSE)

get_pkgs_installation_code_cran(x)

get_pkgs_installation_code_github(x)

get_pkgs_installation_code_other(x)
}
\arguments{
\item{list_name}{(character) The name of the list with required R packages.
E.g., "mini", "Rcmdr", etc.}

\item{use_local_list}{(logical) If \code{TRUE}, the list, which is in the folder
of package \pkg{bio} ("local list"), is used. If \code{FALSE}, the lists
on "GitHub" repository of the package is used. It is recommended
using the online version of the list, as it may contain more recent
changes.}

\item{include}{(character) Which packages from the list (indicated in \code{list_name})
must be included in the results.
One of:
- \code{outdated} (default): only the packages that need are not installed
or do not have a minimum required version installed.
- \code{missing }: only the packages that are not installed to be installed.
- \code{always }: all packages.}

\item{show_status}{(character) Which packages should be included in the
package installation status summary. One of "outdated",
"missing", "always" (see \code{include}). Defaults to the value of \code{include}.}

\item{install}{(character) Which packages should be included in the
package installation code. One of "outdated", "missing", "always"
(see \code{include}). Defaults to the value of \code{include}. Sets the default
value for \code{cran}, \code{github}, and \code{elsewhere}.}

\item{cran}{(character) Condition to filter packages that should be
included in code that installs packages from CRAN. One of "outdated",
"missing", "always" (see \code{include}) as well as "newer_on_cran" and
"required" (see below). Defaults to the value of \code{install}.
- \code{newer_on_cran} -- if CRAN version is newer than the installed one
(even if minimum required version is installed).
- \code{required} -- packages that do not have a minimum required version
installed even if the required version is not on CRAN.}

\item{github}{(character) Condition to filter packages that should be
included in code that installs packages from GitHub. One of "outdated",
"missing", "always" (see \code{include}). Defaults to the value of \code{install}.}

\item{elsewhere}{(character) Condition to filter packages that should
be included in code that installs packages from other sources. One of
"outdated", "missing", "always" (see \code{include}). Defaults to the value
of \code{install}.}

\item{x}{Object of interest.}

\item{...}{Arguments to other methods.}
}
\description{
Get package installation status (e.g., if packages of interest are installed
or need to be updated) and package installation code.
}
\examples{
head(get_pkgs_installation_status_local("mini"))
\dontrun{\donttest{

# NOTE: It is not recommended to use the local lists as they might be out of date.
options(bio.use_local_list = TRUE)
list_name <- "mini"

(status_out <- get_pkgs_installation_status("mini"))
get_pkgs_installation_code(status_out)

(status_all <- get_pkgs_installation_status("mini", include = "always"))
get_pkgs_installation_code(status_all)

(status_custom <-
  get_pkgs_installation_status("mini", include = "always", install = "outdated"))
get_pkgs_installation_code(status_custom)

}}
}
\seealso{
Other R-packages-related functions: 
\code{\link{compare_version}()},
\code{\link{get_pkgs_cran_details}()},
\code{\link{get_pkgs_installed}()},
\code{\link{get_pkgs_non_cran_installation_details}()},
\code{\link{get_pkgs_recommended}()},
\code{\link{get_pkgs_req_version}()}
}
\concept{R-packages-related functions}
