% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_installed}
\alias{get_pkgs_installed}
\title{List packages installed on this computer.}
\usage{
get_pkgs_installed()
}
\value{
Dataframe with columns  "paketas" and "turima_versija".
}
\description{
List packages installed on this computer.
}
\examples{
head(get_pkgs_installed())
}
