% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_installed}
\alias{get_pkgs_installed}
\title{List packages installed on this computer.}
\usage{
get_pkgs_installed()
}
\value{
Data frame with columns \code{"package"} and \code{"current_version"}.
}
\description{
List packages installed on this computer.
}
\examples{

head(get_pkgs_installed())

}
\seealso{
Other R-packages-related functions: 
\code{\link{compare_version}()},
\code{\link{get_pkgs_cran_details}()},
\code{\link{get_pkgs_installation_status_local}()},
\code{\link{get_pkgs_non_cran_installation_details}()},
\code{\link{get_pkgs_recommended}()},
\code{\link{get_pkgs_req_version}()}
}
\concept{R-packages-related functions}
