% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_non_cran_installation_details}
\alias{get_pkgs_non_cran_installation_details}
\title{Get details about non-CRAN package instalation.}
\usage{
get_pkgs_non_cran_installation_details(
  file = "install-r/pkgs-install-from.txt"
)
}
\value{
Dataframe with columns:
\itemize{
\item \code{"package"} package name,
\item \code{"install_from"} source to install from. Currently supported values
are "code" and "github"
\item \code{"details"} either installation code of repository name
\item \code{"notes"}  notes for user.
}
}
\description{
Get installation code of packages that either should be installed not from
CRAN or a modified code shold be installed.
}
\details{
If the file has column \code{ignore} with value \code{TRUE}, the line is not included
in the output.
}
\examples{
\dontrun{\donttest{

head(get_pkgs_non_cran_installation_details())

head(get_pkgs_installation_code())

}}
}
\seealso{
Other R-packages-related functions: 
\code{\link{compare_version}()},
\code{\link{get_pkgs_cran_details}()},
\code{\link{get_pkgs_instalation_status}()},
\code{\link{get_pkgs_installed}()},
\code{\link{get_pkgs_recommended}()},
\code{\link{get_pkgs_req_version}()}
}
\concept{R-packages-related functions}
