% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_recommended}
\alias{get_pkgs_recommended}
\title{List of packages of interest.}
\usage{
get_pkgs_recommended(
  list_name,
  use_local_list = getOption("bio.use_local_list", FALSE)
)
}
\arguments{
\item{list_name}{(character) The name of the list with required R packages.
E.g., "mini", "Rcmdr", etc.}

\item{use_local_list}{(logical) If \code{TRUE}, the list, which is in the folder
of package \pkg{bio} ("local list"), is used. If \code{FALSE}, the lists
on "GitHub" repository of the package is used. It is recommended
using the online version of the list, as it may contain more recent
changes.}
}
\value{
Data frame with column \code{"package"}.
}
\description{
List of packages of interest.
}
\examples{
# NOTE: It is not recommended to use the local lists as they might be out of date.
# Here it is used for testing purposes only.
options(bio.use_local_list = TRUE)

head(get_pkgs_recommended("mini"))

}
\seealso{
Other R-packages-related functions: 
\code{\link{compare_version}()},
\code{\link{get_pkgs_cran_details}()},
\code{\link{get_pkgs_installation_status_local}()},
\code{\link{get_pkgs_installed}()},
\code{\link{get_pkgs_non_cran_installation_details}()},
\code{\link{get_pkgs_req_version}()}
}
\concept{R-packages-related functions}
