% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_req_version}
\alias{get_pkgs_req_version}
\title{Get required version of packages.}
\usage{
get_pkgs_req_version(file = url_bio("pkgs-required-version.txt"))
}
\value{
Dataframe with columns  "package" and "required_version".
}
\description{
Get required version of packages from a list in a file.
}
\examples{
\dontrun{\donttest{

head(get_pkgs_req_version())

}}
}
\seealso{
Other R-packages-related functions: 
\code{\link{compare_version}()},
\code{\link{get_pkgs_cran_details}()},
\code{\link{get_pkgs_instalation_status}()},
\code{\link{get_pkgs_installed}()},
\code{\link{get_pkgs_non_cran_installation_details}()},
\code{\link{get_pkgs_recommended}()}
}
\concept{R-packages-related functions}
