% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{get_pkgs_req_version}
\alias{get_pkgs_req_version}
\title{Get required version of packages.}
\usage{
get_pkgs_req_version(file = "pkgs-required-version")
}
\value{
Dataframe with columns  "paketas" and "reikiama_versija".
}
\description{
Get required version of packages from a list in a file.
}
