% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{is_pkg_installed}
\alias{is_pkg_installed}
\title{Check if packae is installed}
\usage{
is_pkg_installed(pkgs)
}
\arguments{
\item{pkgs}{(character) A list of installed packages.}
}
\value{
A logical vector for each input element.
}
\description{
Check if packae is installed
}
\examples{

is_pkg_installed("bio")

is_pkg_installed(c("bio", "utils", "grugru"))

}
\concept{packages}
