% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages--check.R, R/packages--find.R
\name{list_pkgs_used_in_dir}
\alias{list_pkgs_used_in_dir}
\alias{list_pkgs_used_in_files}
\title{List packages used in file(s)}
\usage{
list_pkgs_used_in_dir(path = ".", regexp = "(?i)[.](rmd|r)$", ...)

list_pkgs_used_in_files(files)

list_pkgs_used_in_dir(path = ".", regexp = "(?i)[.](rmd|r)$", ...)

list_pkgs_used_in_files(files)
}
\arguments{
\item{path}{Path to directory with R and R Markdown files. Defaults to \code{"."}.}

\item{regexp}{Regular expression to filter file names. Defaults to R and Rmd
files.}

\item{...}{Further arguments to \link[fs:dir_ls]{fs::dir_ls}.}

\item{files}{Path to R or Rmd files}
}
\value{
Character vector with package names.

Character vector with package names.
}
\description{
These functions search for patterns, such as \code{library(pkg)}, \code{require(pkg)},
\verb{pkg::function()}, or \code{data(package = "pkg")} and return package names from
these expressions.

These functions search for patterns, such as \code{library(pkg)}, \code{require(pkg)},
\verb{pkg::function()}, or \code{data(package = "pkg")} and return package names from
these expressions.
}
