% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{open_files}
\alias{open_files}
\alias{open_r_user_dir}
\alias{open_rs_config_dir}
\alias{open_rs_1.3_user_settings_dir}
\alias{get_path_rs_desktop_user_settings_dir}
\alias{open_rs_desktop_user_settings_dir}
\alias{get_path_rs_user_settings}
\alias{open_rs_user_settings}
\alias{open_rs_desktop_config_dir}
\alias{open_rs_snippets_dir}
\alias{open_rs_keybindings_dir}
\alias{get_path_r_environ}
\alias{open_r_environ}
\title{Open files and directories}
\usage{
open_r_user_dir()

open_rs_config_dir()

open_rs_1.3_user_settings_dir()

get_path_rs_desktop_user_settings_dir()

open_rs_desktop_user_settings_dir()

get_path_rs_user_settings(which = "current", rstudio_version = "auto")

open_rs_user_settings()

open_rs_desktop_config_dir()

open_rs_snippets_dir()

open_rs_keybindings_dir()

get_path_r_environ(scope = c("user", "project"))

open_r_environ()
}
\arguments{
\item{which}{(character) type of settings: "current", "bio-default".}

\item{rstudio_version}{Numeric version of RStudio (the output is different
for RStudio < 1.3 and 1.3 or newer series). If \code{"auto"}, the version
of RStudio will be determined automatically.}

\item{scope}{(character) One of "user" or "project".}
}
\description{
Open RStudio related files and directories.

\itemize{
\item \code{get_path_r_environ()}-- gets path to file \code{.Renviron} with R environment
variables.
}
}
\examples{
\dontrun{\donttest{

get_path_rs_user_settings()

get_path_rs_user_settings("bio-default")

}}
get_path_r_environ()

}
\seealso{
\itemize{
\item \code{\link[fs:file_show]{fs::file_show()}}, \code{\link[=browseURL]{browseURL()}},
\item \code{\link[rstudioapi:navigateToFile]{rstudioapi::navigateToFile()}},
\item \code{\link[utils:file.edit]{utils::file.edit()}}
}
}
\concept{paths and dirs}
