% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{open_files}
\alias{open_files}
\alias{open_desktop}
\alias{open_r_user_dir}
\alias{open_rstudio_config_dir}
\alias{open_rstudio_internal_state_dir}
\alias{open_rstudio_snippets_dir}
\alias{open_rstudio_keybindings_dir}
\alias{get_path_rstudio_config_file}
\alias{open_rstudio_config_file}
\alias{get_path_r_environ}
\alias{open_r_environ}
\title{Open files and directories}
\usage{
open_desktop()

open_r_user_dir()

open_rstudio_config_dir()

open_rstudio_internal_state_dir()

open_rstudio_snippets_dir()

open_rstudio_keybindings_dir()

get_path_rstudio_config_file(which = "current")

open_rstudio_config_file()

get_path_r_environ(scope = c("user", "project"))

open_r_environ()
}
\arguments{
\item{which}{(character) type of settings: "current", "bio-default".}

\item{scope}{(character) One of "user" or "project".}
}
\description{
Open RStudio related files and directories.

\itemize{
\item \code{get_path_r_environ()}-- gets path to file \code{.Renviron} with R environment
variables.
}
}
\examples{
\dontrun{\donttest{

get_path_rstudio_config_file()

get_path_rstudio_config_file("bio-default")

}}
get_path_r_environ()

}
\seealso{
\itemize{
\item \code{\link[fs:file_show]{fs::file_show()}}, \code{\link[=browseURL]{browseURL()}},
\item \code{\link[rstudioapi:navigateToFile]{rstudioapi::navigateToFile()}},
\item \code{\link[utils:file.edit]{utils::file.edit()}}
}
}
\concept{paths and dirs}
