% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_in_rstudio.R
\name{open_in_rstudio}
\alias{open_in_rstudio}
\title{Open file in RStudio}
\usage{
open_in_rstudio(path, ...)
}
\arguments{
\item{path}{(sting) Path to file.}

\item{...}{Further arguments to \code{\link[rstudioapi:navigateToFile]{rstudioapi::navigateToFile()}}.}
}
\description{
Function tries opening a file in RStudio.
}
\seealso{
\itemize{
\item \code{\link[rstudioapi:navigateToFile]{rstudioapi::navigateToFile()}},
\item \code{\link[fs:file_show]{fs::file_show()}}, \code{\link[=browseURL]{browseURL()}},
\item \code{\link[utils:file.edit]{utils::file.edit()}}
\item \code{\link[usethis:edit_file]{usethis::edit_file()}}
}
}
\concept{open}
\concept{open files}
\concept{paths and dirs}
