% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{open_project}
\alias{open_project}
\alias{get_path_personal_proj_list}
\alias{open_personal_proj_list}
\alias{update_personal_proj_list}
\alias{open_project_personal}
\alias{get_projs_all}
\title{Open RStudio project.}
\usage{
open_project(
  name = NULL,
  new_session = FALSE,
  proj_list = NULL,
  proj_list_path = get_path_recent_proj_list(),
  only_existing = TRUE
)

get_path_personal_proj_list()

open_personal_proj_list()

update_personal_proj_list()

open_project_personal(name = NULL, new_session = FALSE, ...)

get_projs_all()
}
\arguments{
\item{name}{(string|\code{NULL}) The name of the project or \code{NULL} to choose
a project interactively.}

\item{new_session}{(logical) should the project be opened in a new session,
or should the current RStudio session switch to that project? Note
that \code{TRUE} values are only supported with RStudio Desktop and RStudio
Server Pro.}

\item{proj_list}{(data frame) The result of \code{\link[=read_projects]{read_projects()}} or \code{NULL}.}

\item{proj_list_path}{(string) The path to the file with the list of project
paths. If \code{proj_list} is not \code{NULL}, then \code{proj_list_path} is ignored.}

\item{only_existing}{(logical) If \code{TRUE}, non-existing projects and projects
with broken paths are removed from the list of choices.}
}
\value{
Opens the indicated project.
}
\description{
Open RStudio project by name or interactively.
}
\examples{
\dontrun{\donttest{

open_project()

open_project("bs")

open_project("R-2019-project")

}}
}
\seealso{
\itemize{
\item \code{\link[rstudioapi:openProject]{rstudioapi::openProject()}}
\item \code{\link[rstudioapi:initializeProject]{rstudioapi::initializeProject()}}
}
}
