% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{open_project}
\alias{open_project}
\title{Open RStudio Project}
\usage{
open_project(
  pattern = NULL,
  new_session = TRUE,
  proj_list = NULL,
  proj_list_path = NULL,
  only_available = TRUE,
  name = NULL,
  negate = FALSE
)
}
\arguments{
\item{pattern}{(character) regular expression to narrow down the list of
possible options.}

\item{new_session}{(logical) should the project be opened in a new session,
or should the current RStudio session switch to that project? Note
that \code{TRUE} values are only supported with RStudio Desktop and RStudio
Server Pro.}

\item{proj_list}{(data frame) The result of \code{\link[=read_projects]{read_projects()}} or \code{NULL}.}

\item{proj_list_path}{(string) The path to the file with the list of project
paths. If \code{proj_list} is not \code{NULL}, then \code{proj_list_path} is ignored.}

\item{only_available}{(logical) If \code{TRUE}, non-existing projects and projects
with broken paths are removed from the list of choices.}

\item{name}{(string|\code{NULL}) The name of the project or \code{NULL} to choose
a project interactively.}

\item{negate}{(logical) If \code{TRUE}, then the options defined by  \code{pattern} are
excluded.}

\item{...}{Further arguments.}
}
\value{
Opens the indicated project.
}
\description{
Open RStudio project by name or interactively. The projects list is read from
files that contain projeect lists.
}
\examples{
\dontrun{\donttest{

open_project()

open_project("bio")

open_project("R-2019-project")

}}
}
\seealso{
\itemize{
\item \code{\link[rstudioapi:projects]{rstudioapi::openProject()}}
\item \code{\link[rstudioapi:projects]{rstudioapi::initializeProject()}}
}
}
\concept{rstudio projects}
