% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages--check.R
\name{pkg_list_archived_versions}
\alias{pkg_list_archived_versions}
\title{Get previous package versions available on CRAN}
\usage{
pkg_list_archived_versions(package)
}
\arguments{
\item{package}{(character)
Package name.}
}
\value{
Vector with version numbers (the current version is not present).
}
\description{
Function to scrape the CRAN website and retrieve archived (old) package versions
}
\examples{
list_archived_package_versions("ggplot2")

list_archived_package_versions("none")
}
