% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{project-lists}
\alias{project-lists}
\alias{get_path_recent_proj_list}
\alias{get_path_personal_proj_list}
\alias{open_recent_proj_list}
\alias{open_personal_proj_list}
\alias{update_personal_proj_list}
\title{Manage Project Lists}
\usage{
get_path_recent_proj_list()

get_path_personal_proj_list(create = FALSE)

open_recent_proj_list()

open_personal_proj_list()

update_personal_proj_list()
}
\arguments{
\item{create}{(logical) If \code{TRUE} and file does not exist, the file is created.}
}
\description{
Manage project lists.

\itemize{
\item \code{get_path_recent_proj_list()} -- gets path to the file with the list of
recent RStudio projects.
\item \code{get_path_personal_proj_list()} -- gets path to the file with the list of
personal RStudio projects.
}

\itemize{
\item \code{open_recent_proj_list()} -- opens the file with the list of
recent RStudio projects.
}
}
\examples{
\dontrun{\donttest{
get_path_recent_proj_list()
get_path_personal_proj_list()
}}

}
\concept{rstudio projects}
