% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{read_projects}
\alias{read_projects}
\alias{projects}
\alias{parse_proj_path}
\alias{get_path_recent_proj_list}
\alias{get_projs_recent}
\alias{get_proj_names}
\alias{open_recent_proj_list}
\title{Project paths.}
\usage{
read_projects(file, sort_by = FALSE)

parse_proj_path(proj)

get_path_recent_proj_list()

get_projs_recent(sort_by = FALSE)

get_proj_names(file = get_path_recent_proj_list(), sort_by = FALSE)

open_recent_proj_list()
}
\arguments{
\item{file}{(character) Path to file with RStudio projects.}

\item{sort_by}{(\code{"name"}|\code{"path"}|\code{\link{FALSE}})}

\item{proj}{(character) Path to *.Rproj file.}
}
\value{
\itemize{
\item \code{parse_proj_path()} returns data frame with columns \code{name} (with project
name, i.e., folder name) and \code{path} to *.Rproj file.
}
}
\description{
\itemize{
\item \code{read_projects()} - reads file with projects and list their names and paths.
}

\itemize{
\item \code{get_path_recent_proj_list()} -- gets path to the file with the list of
recent RStudio projects.
}

\itemize{
\item \code{get_projs_recent()} -- lists recent RStudio projects (into a data
frame).
}

\itemize{
\item \code{get_proj_names()} -- lists RStudio projects names (as a character
vector)
}

\itemize{
\item \code{open_recent_proj_list()} -- opens the file with the list of
recent RStudio projects.
}
}
\examples{
\dontrun{\donttest{
parse_proj_path("C:/data/analysis/proj.Rproj")
}}
\dontrun{\donttest{
get_path_recent_proj_list()
}}

\dontrun{\donttest{
head(get_proj_names())
}}
}
