% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{rproj_read_projects}
\alias{rproj_read_projects}
\alias{projects}
\alias{rproj_path_to_df}
\alias{rproj_get_path_rs_recent_proj_list}
\alias{rproj_get_current_projects}
\alias{rproj_get_proj_names}
\alias{rproj_open_rs_recent_proj_list}
\title{Project paths.}
\usage{
rproj_read_projects(file, sort_by = FALSE)

rproj_path_to_df(proj)

rproj_get_path_rs_recent_proj_list()

rproj_get_current_projects(sort_by = FALSE)

rproj_get_proj_names(
  file = rproj_get_path_rs_recent_proj_list(),
  sort_by = FALSE
)

rproj_open_rs_recent_proj_list()
}
\arguments{
\item{file}{(character) Path to file with RStudio projects.}

\item{sort_by}{(\code{"name"}|\code{"path"}|\code{\link{FALSE}})}

\item{proj}{(character) Path to *.Rproj file.}
}
\value{
\itemize{
\item \code{rproj_path_to_df()} returns data frame with columns \code{name} (with project
name, i.e., folder name) and \code{path} to *.Rproj file.
}
}
\description{
\itemize{
\item \code{rproj_read_projects()} - reads file with projects and list their names and paths.
}

\itemize{
\item \code{rproj_get_path_rs_recent_proj_list()} -- gets path to the file with the list of
recent RStudio projects.
}

\itemize{
\item \code{rproj_get_current_projects()} -- lists current RStudio projects (in a data
frame).
}

\itemize{
\item \code{rproj_get_proj_names()} -- lists RStudio projects names (as a character
vector)
}

\itemize{
\item \code{rproj_open_rs_recent_proj_list()} -- opens the file with the list of
recent RStudio projects.
}
}
\examples{
\dontrun{\donttest{
rproj_path_to_df("C:/data/analysis/proj.Rproj")
}}
\dontrun{\donttest{
rproj_get_path_rs_recent_proj_list()
}}

\dontrun{\donttest{
head(rproj_get_proj_names())
}}
}
