% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects}
\alias{projects}
\alias{read_projects}
\alias{get_path_recent_proj_list}
\alias{get_path_personal_proj_list}
\alias{get_projs_recent}
\alias{get_projs_personal}
\alias{get_projs_all}
\alias{get_proj_names}
\title{Manage RStudio projects.}
\usage{
read_projects(file, sort_by = FALSE)

get_path_recent_proj_list()

get_path_personal_proj_list(create = FALSE)

get_projs_recent(sort_by = FALSE)

get_projs_personal(sort_by = FALSE)

get_projs_all()

get_proj_names(file = get_path_recent_proj_list(), sort_by = FALSE)
}
\arguments{
\item{file}{(character) Path to file with RStudio projects.}

\item{sort_by}{(\code{"name"}|\code{"path"}|\code{\link{FALSE}})}

\item{create}{(ligical) If \code{TRUE} and file does not exist, the file is created.}
}
\description{
\itemize{
\item \code{read_projects()} - reads file with projects and list their names and paths.
}

\itemize{
\item \code{get_path_recent_proj_list()} -- gets path to the file with the list of
recent RStudio projects.
\item \code{get_path_personal_proj_list()} -- gets path to the file with the list of
personal RStudio projects.
}

\itemize{
\item \code{get_projs_recent()} -- lists recent RStudio projects.
\item \code{get_projs_personal()} -- lists personal RStudio projects.
\item \code{get_projs_all()} -- lists both recent and personal RStudio projects.
}

\itemize{
\item \code{get_proj_names()} -- lists RStudio projects names (as a character
vector)
}
}
\examples{
\dontrun{\donttest{
read_projects(get_path_recent_proj_list())
}}

\dontrun{\donttest{
get_path_recent_proj_list()
get_path_personal_proj_list()
}}

\dontrun{\donttest{
get_projs_recent()
get_projs_personal()
get_projs_all()
}}

\dontrun{\donttest{
head(get_proj_names())
}}
}
