% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{open_backup_dir}
\alias{get_path_rstudio_snippets_dir}
\alias{open_rstudio_snippets_dir}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{backup.tools}{\code{\link[backup.tools:get_path_backup_dir]{open_backup_dir}}}

  \item{snippets}{\code{\link[snippets]{get_path_rstudio_snippets_dir}}, \code{\link[snippets]{open_rstudio_snippets_dir}}}
}}

