% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R, R/git.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{selectDirectory}
\alias{edit_git_config}
\alias{use_git_config}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rstudioapi}{\code{\link[rstudioapi]{selectDirectory}}}

  \item{usethis}{\code{\link[usethis]{edit_git_config}}, \code{\link[usethis]{use_git_config}}}
}}

