% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R, R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{edit_git_config}
\alias{use_git_config}
\alias{open_backup_dir}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{backup.tools}{\code{\link[backup.tools:get_path_backup_dir]{open_backup_dir}}}

  \item{usethis}{\code{\link[usethis:edit]{edit_git_config}}, \code{\link[usethis]{use_git_config}}}
}}

