% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED-functions.R
\name{reload_rstudio}
\alias{reload_rstudio}
\title{DEPRECATED. Use \code{rstudio_reload()} instead}
\usage{
reload_rstudio(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_reload]{rstudio_reload()}}}
}
\description{
DO NOT USE THIS FUNCTION!
This function was replaced by \code{\link[=rstudio_reload]{rstudio_reload()}} and will be removed soon.
}
\concept{deprecated}
