% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{reset_rstudio_gmc}
\alias{reset_rstudio_gmc}
\title{DEFUNCT. Use \code{rstudio_reset_gmc()} instead}
\usage{
reset_rstudio_gmc(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_reset_gmc]{rstudio_reset_gmc()}}}
}
\description{
\strong{DO NOT USE THIS FUNCTION!}
This function was replaced by \code{\link[=rstudio_reset_gmc]{rstudio_reset_gmc()}} and will be removed soon.
}
\concept{defunct}
