% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{reset_rstudio_gmc}
\alias{reset_rstudio_gmc}
\title{Reset RStudio state in GMC R209 and clear environment.}
\usage{
reset_rstudio_gmc(..., force_update_dictionaries = FALSE)
}
\arguments{
\item{...}{Further arguments for advanced users only.}

\item{force_update_dictionaries}{(logical) If \code{TRUE}, the dictionaries are
forced to be downloaded/updated.}
}
\description{
This function:
\enumerate{
\item Resets RStudio state and user preferences (incl. color scheme)
\item Clears function history, plot history, console, recent project list, etc.
\item Closes unnecessary windows
\item Resets custom keybindings tto "bio-default"
\item Resets R Markdown and R snippets to defaults in package "snippets"
\item Creates folder "~/R/Darbinis" and starts using it as working directory
when no project is used.
\item Clears (if possible)/creates folder "BS-2020" on Desktop.
}
}
\details{
The function works only in GMC R209, GMC C255 and ITPC EC-108 classrooms.
}
\examples{
\dontrun{\donttest{

bio::reset_rstudio_gmc()

}}
}
