% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{reset_rstudio_keybindings}
\alias{reset_rstudio_keybindings}
\title{Set RStudio keybindings}
\usage{
reset_rstudio_keybindings(to, backup = TRUE)
}
\arguments{
\item{to}{(string) The type of keybindings.
Currently available value is \code{"bio-default"}.}

\item{backup}{(logical) If \code{TRUE}, a back-up copy of files with current
keybindings will be created.}
}
\description{
Set RStudio keybindings
}
\examples{
\dontrun{\dontest{

bio::reset_rstudio_keybindings(to = "bio-default")
bio::reload_rstudio()

}}
}
\concept{r and rstudio settings}
