% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{reset_rstudio_keybindings}
\alias{reset_rstudio_keybindings}
\title{DEPRECATED. Use \code{rstudio_reset_keybindings()} instead}
\usage{
reset_rstudio_keybindings(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=rstudio_reset_keybindings]{rstudio_reset_keybindings()}}}
}
\description{
\strong{DO NOT USE THIS FUNCTION!}
This function was replaced by \code{\link[=rstudio_reset_keybindings]{rstudio_reset_keybindings()}} and will be removed soon.
}
\concept{deprecated}
