% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restart_reload.R
\name{restart-reload}
\alias{restart-reload}
\alias{rstudio_restart_r}
\alias{rstudio_reload_ui}
\title{Functions to Restart R and Reload RStudio}
\usage{
rstudio_restart_r()

rstudio_reload_ui()
}
\description{
\itemize{
\item \code{rstudio_restart_r()} restarts R session in RStudio.
\item \code{rstudio_reload()} reloads RStudio without closing it.
}
}
\concept{utilities}
