% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{restart-reload}
\alias{restart-reload}
\alias{rstudio_restart_r}
\alias{rstudio_reload}
\title{Functions that to restart or reload programs}
\usage{
rstudio_restart_r()

rstudio_reload()
}
\description{
\code{rstudio_restart_r()} restarts R session in RStudio.
\code{rstudio_reload()} reloads RStudio without closing it.
}
\concept{utilities}
