% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{restart-reload}
\alias{restart-reload}
\alias{restart_r}
\alias{reload_rstudio}
\title{Functions that to restart or reload programs}
\usage{
restart_r()

reload_rstudio()
}
\description{
\code{restart_r()} restarts R session in RStudio.
\code{reload_rstudio()} reloads RStudio without closing it.
}
