% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{RStudio-config-dir}
\alias{RStudio-config-dir}
\alias{get_path_r_user_dir}
\alias{get_path_rs_config_dir}
\alias{get_path_user_settings_dir_rs_1.3}
\alias{get_path_rstudio_config_dir}
\alias{get_path_rs_snippets_dir}
\alias{get_path_rs_keybindings_dir}
\title{Directories of R and RStudio (desktop) settings, preferences, etc.}
\usage{
get_path_r_user_dir(...)

get_path_rs_config_dir(...)

get_path_user_settings_dir_rs_1.3(...)

get_path_rstudio_config_dir(..., .check = FALSE, rstudio_version = "auto")

get_path_rs_snippets_dir(rstudio_version = "auto")

get_path_rs_keybindings_dir(rstudio_version = "auto")
}
\arguments{
\item{...}{(character) Parts of the path. Path to sub-directories.}

\item{.check}{(logical) If \code{TRUE}, additionally checks for path existance.}

\item{rstudio_version}{Numeric version of RStudio (the output is different
for RStudio < 1.3 and 1.3 or newer series). If \code{"auto"}, the version
of RStudio will be determined automatically.}
}
\value{
\itemize{
\item \code{get_path_rs_config_dir()} (string) path, if it exists, or error.
}

(string) path to RStudio desktop user settings directory.
When \code{.check = TRUE}, renturns error, if the path does not exist.
}
\description{
\itemize{
\item \code{get_path_r_user_dir()}-- gets path to the main R user directory.
}

\itemize{
\item \code{get_path_rs_config_dir()} -- gets path to RStudio configuration directory
or its sub-directories.
}

\itemize{
\item `get_path_rstudio_config_dir()`` - gets path to directory (and its
sub-directories) of RStudio desktop user interface settings.
}
}
\examples{
get_path_r_user_dir()

\dontrun{\donttest{
get_path_rs_config_dir()
}}
\dontrun{\donttest{
get_path_rstudio_config_dir()

get_path_rstudio_config_dir("dictionaries")
}}
\dontrun{\donttest{
get_path_rs_snippets_dir()
}}

get_path_rs_snippets_dir(rstudio_version = "1.4.1103")

\dontrun{\donttest{
get_path_rs_keybindings_dir()
}}

get_path_rs_keybindings_dir(rstudio_version = "1.4.1103")

}
\seealso{
\itemize{
\item \code{get_path_rstudio_config_dir()}:
https://support.rstudio.com/hc/en-us/articles/200534577-Resetting-RStudio-Desktop-s-State
}
}
\concept{paths and dirs}
