% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{RStudio-config-file}
\alias{RStudio-config-file}
\alias{get_path_rstudio_config_file}
\alias{open_rstudio_config_file}
\title{Manage RStudio Configuration (Preferences) File}
\usage{
get_path_rstudio_config_file(which = "current")

open_rstudio_config_file(which = "current")
}
\arguments{
\item{which}{(character) type of settings:
\itemize{
\item "current": file with current RStudio settings (that differ from the defaults);
\item "bio-default": file with setting from "bio-default" list (except theme);
\item "rstudio-default": file with most of default RStudio settings listed at
https://docs.rstudio.com/ide/server-pro/session_user_settings/session_user_settings.html (downloaded on 2022-08-03).
}}
}
\description{
Manage file with RStudio configuration (user preferences).
}
\examples{
\dontrun{\donttest{

get_path_rstudio_config_file()

get_path_rstudio_config_file("bio-default")

}}
}
\seealso{
\itemize{
\item \code{\link[=get_path_rstudio_config_dir]{get_path_rstudio_config_dir()}}
}
}
\concept{paths and dirs}
