% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{RStudio-config-file}
\alias{RStudio-config-file}
\alias{get_path_rstudio_config_file}
\alias{open_rstudio_config_file}
\title{Manage RStudio Configuration (Preferences) File}
\usage{
get_path_rstudio_config_file(which = "current")

open_rstudio_config_file()
}
\arguments{
\item{which}{(character) type of settings: "current", "bio-default".}
}
\description{
Manage file with RStudio configuration (user preferences).
}
\examples{
\dontrun{\donttest{

get_path_rstudio_config_file()

get_path_rstudio_config_file("bio-default")

}}
}
\seealso{
\itemize{
\item \code{\link[=get_path_rstudio_config_dir]{get_path_rstudio_config_dir()}}
}
}
