% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{RStudio-dictionaries}
\alias{RStudio-dictionaries}
\alias{open_rstudio_system_dictionaries_dir}
\alias{open_rstudio_user_dictionaries_dir}
\alias{open_rstudio_internal_dictionaries_dir}
\title{RStudio Dictionaries}
\usage{
open_rstudio_system_dictionaries_dir()

open_rstudio_user_dictionaries_dir()

open_rstudio_internal_dictionaries_dir()
}
\value{
String with path.
}
\description{
Functions to work with RStudio dictionaries.

\itemize{
\item \code{\link[rstudioapi:dictionariesPath]{rstudioapi::dictionariesPath()}}
\item \code{open_rstudio_system_dictionaries_dir()}
}

\itemize{
\item \code{\link[rstudioapi:userDictionariesPath]{rstudioapi::userDictionariesPath()}}
\item \code{open_rstudio_user_dictionaries_dir()}
}

\itemize{
\item \code{\link[rstudioapi:dictionariesPath]{rstudioapi::dictionariesPath()}}
\item \code{open_rstudio_internal_dictionaries_dir()}
}
}
\examples{
\dontrun{\donttest{

rstudioapi::dictionariesPath()

rstudioapi::userDictionariesPath()

get_path_rstudio_config_dir("dictionaries")

}}
}
\seealso{
\link[rstudioapi:dictionaries]{rstudioapi::dictionaries}
}
\concept{dictionaries}
\concept{paths and dirs}
