% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{RStudio-dictionaries}
\alias{RStudio-dictionaries}
\alias{get_path_saved_words_dictionary}
\alias{get_path_rs_saved_words_dictionary}
\alias{open_rs_saved_words_dictionary}
\alias{get_path_rs_system_dictionaries_dir}
\alias{open_rs_system_dictionaries_dir}
\alias{open_rs_dictionaries_dir}
\alias{open_user_dictionaries_dir}
\title{RStudio dictionaries}
\usage{
get_path_saved_words_dictionary()

get_path_rs_saved_words_dictionary()

open_rs_saved_words_dictionary()

get_path_rs_system_dictionaries_dir()

open_rs_system_dictionaries_dir()

open_rs_dictionaries_dir()

open_user_dictionaries_dir()
}
\value{
String with path.
}
\description{
Functions to work with RStudio dictionaries.

\itemize{
\item \code{get_path_rs_saved_words_dictionary()} -- Path to dictionary of words
saved by user
\item \code{open_rs_saved_words_dictionary()}
}

\itemize{
\item \code{get_path_rs_system_dictionaries_dir()}
\item \code{open_rs_system_dictionaries_dir()}
}

\itemize{
\item \code{\link[rstudioapi:dictionaries]{rstudioapi::dictionariesPath()}}
\item \code{open_rs_dictionaries_dir()}
}

\itemize{
\item \code{\link[rstudioapi:dictionaries]{rstudioapi::userDictionariesPath()}}
\item \code{open_user_dictionaries_dir()}
}
}
\examples{
if (FALSE) {

rstudioapi::dictionariesPath()

rstudioapi::userDictionariesPath()

get_path_rs_system_dictionaries_dir()
}

get_path_rs_desktop_config_dir("dictionaries")

get_path_saved_words_dictionary()


}
\seealso{
\link[rstudioapi:dictionaries]{rstudioapi::dictionaries}
}
\concept{dictionaries}
\concept{paths and dirs}
