% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{RStudio-related-dirs}
\alias{RStudio-related-dirs}
\alias{get_path_rstudio_config_dir}
\alias{get_path_rstudio_internal_state_dir}
\alias{get_path_rstudio_keybindings_dir}
\alias{open_rstudio_config_dir}
\alias{open_rstudio_internal_state_dir}
\alias{open_rstudio_keybindings_dir}
\title{Directories of RStudio-Related Files}
\usage{
get_path_rstudio_config_dir(..., .check = FALSE)

get_path_rstudio_internal_state_dir(..., .check = FALSE)

get_path_rstudio_keybindings_dir()

open_rstudio_config_dir()

open_rstudio_internal_state_dir()

open_rstudio_keybindings_dir()
}
\arguments{
\item{...}{(character) Parts of the path. Path to sub-directories.}

\item{.check}{(logical) If \code{TRUE}, additionally checks for path existence.}
}
\value{
(string) path to RStudio configuration directory.
When \code{.check = TRUE}, returns an error if the path does not exist.
}
\description{
Directories of RStudio (desktop) settings, preferences and other files.

\itemize{
\item `get_path_rstudio_config_dir()`` - gets path to RStudio configuration
directory (and its sub-directories).
}
}
\examples{
if (interactive()) {
  get_path_rstudio_config_dir()

  get_path_rstudio_config_dir("dictionaries")
}
if (interactive()) {
  get_path_rstudio_keybindings_dir()
}
}
\seealso{
\itemize{
\item \code{\link[fs:file_show]{fs::file_show()}}, \code{\link[=browseURL]{browseURL()}},
\item \code{\link[rstudioapi:navigateToFile]{rstudioapi::navigateToFile()}},
\item \code{\link[utils:file.edit]{utils::file.edit()}}
}

\itemize{
\item \code{get_path_rstudio_config_dir()}:
https://support.posit.co/hc/en-us/articles/200534577-Resetting-RStudio-Desktop-s-State
}
}
\concept{paths and dirs}
