% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths-and-files.R
\name{RStudio-related-dirs}
\alias{RStudio-related-dirs}
\alias{get_path_r_user_dir}
\alias{get_path_rstudio_config_dir}
\alias{get_path_rstudio_internal_state_dir}
\alias{get_path_rstudio_snippets_dir}
\alias{get_path_rstudio_keybindings_dir}
\alias{open_r_user_dir}
\alias{open_rstudio_config_dir}
\alias{open_rstudio_internal_state_dir}
\alias{open_rstudio_snippets_dir}
\alias{open_rstudio_keybindings_dir}
\title{Directories of R and RStudio Related Files}
\usage{
get_path_r_user_dir(...)

get_path_rstudio_config_dir(..., .check = FALSE)

get_path_rstudio_internal_state_dir(..., .check = FALSE)

get_path_rstudio_snippets_dir()

get_path_rstudio_keybindings_dir()

open_r_user_dir()

open_rstudio_config_dir()

open_rstudio_internal_state_dir()

open_rstudio_snippets_dir()

open_rstudio_keybindings_dir()
}
\arguments{
\item{...}{(character) Parts of the path. Path to sub-directories.}

\item{.check}{(logical) If \code{TRUE}, additionally checks for path existance.}
}
\value{
(string) path to RStudio configuration directory.
When \code{.check = TRUE}, renturns error, if the path does not exist.
}
\description{
Directories of R and RStudio (desktop) settings, preferences and other files.
\itemize{
\item \code{get_path_r_user_dir()}-- gets path to the main R user directory.
}

\itemize{
\item `get_path_rstudio_config_dir()`` - gets path to RStudio configuration
directory (and its sub-directories).
}
}
\examples{
get_path_r_user_dir()

\dontrun{\donttest{
get_path_rstudio_config_dir()

get_path_rstudio_config_dir("dictionaries")
}}
\dontrun{\donttest{
get_path_rstudio_snippets_dir()
}}
\dontrun{\donttest{
get_path_rstudio_keybindings_dir()
}}
}
\seealso{
\itemize{
\item \code{get_path_rstudio_config_dir()}:
https://support.rstudio.com/hc/en-us/articles/200534577-Resetting-RStudio-Desktop-s-State
}
}
\concept{paths and dirs}
