% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings--preferences.R
\name{RStudio-settings}
\alias{RStudio-settings}
\alias{rstudio_reset_user_settings}
\title{Reset RStudio settings}
\usage{
rstudio_reset_user_settings(to, backup = TRUE, ask = TRUE)
}
\arguments{
\item{to}{The name of pre-defined set of RStudio settings/preferences.
Options: "rstudio-default",
"bio-default",
"bio-dark-blue",
"bio-black".}

\item{backup}{(logical)
If \code{TRUE}, a backup copy of files with settings is created.}

\item{ask}{(logical)
If \code{TRUE}, additional confirmation to reset settings is required.}
}
\description{
Reset RStudio to use predefined set of settings/preferences.
Correctly works only with RStudio 1.3 or newer.
Recommended to use with RStudio 2022.07.1 or newer.
}
\details{
Settings that can be used in \code{rstudio-prefs.json} file:
https://docs.posit.co/ide/server-pro/session-user-settings.html
}
\examples{
if (interactive()) {

  rstudio_reset_user_settings(to = "rstudio-default")
  rstudio_reset_user_settings(to = "bio-default")
  rstudio_reset_user_settings(to = "bio-dark-blue")
  rstudio_reset_user_settings(to = "bio-black")

}
}
\seealso{
\code{\link[=get_path_rstudio_config_file]{get_path_rstudio_config_file()}}

On \href{https://support.posit.co/hc/en-us/articles/200549016-Customizing-the-RStudio-IDE}{Customizing RStudio} using point-and-click method.

On \href{https://www.rstudio.com/blog/rstudio-1-3-preview-configuration/}{Configuration and Settings}.

A list of \href{https://docs.posit.co/ide/server-pro/session_user_settings/session_user_settings.html}{Session User Settings} to be used with
\code{\link[rstudioapi:writeRStudioPreference]{rstudioapi::writeRStudioPreference()}}.

On \href{https://docs.posit.co/ide/desktop-pro/settings/settings.html}{RStudio setting locations}.

On \href{https://support.posit.co/hc/en-us/articles/200534577-Resetting-RStudio-Desktop-s-State}{Resetting RStudio Desktop's State}.

StackOverflow threads on export/import RStudio of user preferences:
\itemize{
\item https://stackoverflow.com/a/55940249/4783029
\item https://stackoverflow.com/a/54982341/4783029
}
}
\concept{r and rstudio settings}
