% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings--preferences.R
\name{rstudio_compare_user_settings}
\alias{rstudio_compare_user_settings}
\title{Show differences in sets of settings}
\usage{
rstudio_compare_user_settings(to = "bio-default")
}
\arguments{
\item{to}{One of: "bio-default", "rstudio-default"
(or an unambiguous abbreviation of these).}
}
\value{
Nothing. But prints the set differences between \code{to} list and
current settings. Settings, which are not in \code{to} list, will
not be displayed at all.
}
\description{
Show differences in sets of settings
}
\examples{
\dontrun{\donttest{
rstudio_compare_user_settings(to = "bio-default")
rstudio_compare_user_settings(to = "rstudio-default")
}}
}
