% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{rstudio_reset_layout}
\alias{rstudio_reset_layout}
\title{Reset the RStudio pane layout.}
\usage{
rstudio_reset_layout(rs_layout = "left")
}
\arguments{
\item{rs_layout}{Character scalar: either \code{"left"} or \code{"right"}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Reset the RStudio pane layout.
}
\examples{
if (interactive()) {
  rstudio_reset_layout("left")
}
}
\keyword{internal}
