% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{spelling}
\alias{spelling}
\alias{rstudio_download_spellcheck_dictionaries}
\alias{rstudio_delete_spellcheck_dictionaries}
\title{Dictionaries to check spelling}
\usage{
rstudio_download_spellcheck_dictionaries(
  rstudio_version = "auto",
  secure = TRUE
)

rstudio_delete_spellcheck_dictionaries()
}
\arguments{
\item{rstudio_version}{Numeric version of RStudio (the output is different
for RStudio < 1.3 and 1.3 or newer series). If \code{"auto"}, the version
of RStudio will be determined automatically.}

\item{secure}{(logical) If \code{TRUE}, uses "https", if \code{FALSE}, uses "http".}
}
\description{
\code{rstudio_download_spellcheck_dictionaries()}
downloads and updates RStudio (system) spellchecking dictionaries.
\code{rstudio_delete_spellcheck_dictionaries()}
deletes RStudio (system) spellchecking dictionaries.
}
\examples{
\dontrun{\donttest{
rstudio_delete_spellcheck_dictionaries()
rstudio_download_spellcheck_dictionaries()
}}
}
\concept{dictionaries}
\concept{r and rstudio settings}
