% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{spelling}
\alias{spelling}
\alias{update_rs_system_dictionaries}
\alias{delete_rs_system_dictionaries}
\title{Spellchecking dictionaries.}
\usage{
update_rs_system_dictionaries(secure = TRUE)

delete_rs_system_dictionaries()
}
\arguments{
\item{secure}{(logical) If \code{TRUE}, uses "https", if \code{FALSE}, uses "http".}
}
\description{
\code{update_rs_system_dictionaries()} downloads and updates spellchecking
dictionaries.
\code{delete_rs_system_dictionaries()} deletes spellchecking dictionaries.
}
\examples{
if (FALSE) {

delete_rs_system_dictionaries()

update_rs_system_dictionaries()

}
}
