% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{spelling}
\alias{spelling}
\alias{download_rs_system_dictionaries}
\alias{delete_rs_system_dictionaries}
\title{Dictionaries to check spelling.}
\usage{
download_rs_system_dictionaries(secure = TRUE)

delete_rs_system_dictionaries()
}
\arguments{
\item{secure}{(logical) If \code{TRUE}, uses "https", if \code{FALSE}, uses "http".}
}
\description{
\code{download_rs_system_dictionaries()} downloads and updates RStudio
spellchecking dictionaries.
\code{delete_rs_system_dictionaries()} deletes RStudio spellchecking dictionaries.
}
\examples{
if (FALSE) {

delete_rs_system_dictionaries()

download_rs_system_dictionaries()

}
}
