% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{spelling}
\alias{spelling}
\alias{rstudio_download_spellcheck_dictionaries}
\alias{rstudio_install_spellcheck_dictionary_lt}
\alias{rstudio_delete_spellcheck_dictionaries}
\title{Dictionaries to check spelling}
\usage{
rstudio_download_spellcheck_dictionaries(secure = TRUE)

rstudio_install_spellcheck_dictionary_lt()

rstudio_delete_spellcheck_dictionaries(ask = TRUE)
}
\arguments{
\item{secure}{(logical) If \code{TRUE}, uses "https", if \code{FALSE}, uses "http".}

\item{ask}{(logical) If \code{TRUE}, user will have to confirm his/her choice
interactively.}
}
\description{
\itemize{
\item \code{rstudio_download_spellcheck_dictionaries()}
downloads and updates RStudio (system) spellchecking dictionaries.
\item \code{rstudio_download_spellcheck_dictionaries_lt()}
installs improved Lithuanian spellchecking dictionary.
\item \code{rstudio_delete_spellcheck_dictionaries()}
deletes RStudio (system) spellchecking dictionaries.
}
}
\examples{
\dontrun{\donttest{
rstudio_delete_spellcheck_dictionaries()
rstudio_download_spellcheck_dictionaries()
rstudio_install_spellcheck_dictionary_lt()
}}
}
\concept{dictionaries}
\concept{r and rstudio settings}
