% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{spelling}
\alias{spelling}
\alias{download_spellcheck_dictionaries}
\alias{delete_spellcheck_dictionaries}
\title{Dictionaries to check spelling.}
\usage{
download_spellcheck_dictionaries(secure = TRUE)

delete_spellcheck_dictionaries()
}
\arguments{
\item{secure}{(logical) If \code{TRUE}, uses "https", if \code{FALSE}, uses "http".}
}
\description{
\code{download_spellcheck_dictionaries()}
downloads and updates RStudio (system) spellchecking dictionaries.
\code{delete_spellcheck_dictionaries()}
deletes RStudio (system) spellchecking dictionaries.
}
\examples{
if (FALSE) {

  delete_spellcheck_dictionaries()
download_spellcheck_dictionaries()
}
}
