% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_update.R
\name{update_pkg_rcmdr_biostat}
\alias{update_pkg_rcmdr_biostat}
\alias{check_updates_rcmdr_biostat}
\title{Update '\pkg{RcmdrPlugin.biostat}' and related packages}
\usage{
check_updates_rcmdr_biostat(
  show_status = "always",
  install = "outdated",
  upgrade = FALSE,
  ...
)

update_pkg_rcmdr_biostat(upgrade = FALSE, force = FALSE, quiet = TRUE)
}
\arguments{
\item{upgrade}{Upgrade dependencies.
See \code{upgrade} in \code{\link[remotes:install_github]{remotes::install_github()}}.}

\item{force}{(logical) \code{TRUE} or \code{FALSE}.
Force to update.
See \code{force} in \code{\link[remotes:install_github]{remotes::install_github()}}.}

\item{quiet}{(logical) \code{TRUE} or \code{FALSE}.
Should installation messages be printed?
See \code{quiet} in \code{\link[remotes:install_github]{remotes::install_github()}}.}
}
\description{
Update '\pkg{RcmdrPlugin.biostat}' and related packages
}
\examples{
\dontrun{\donttest{

update_pkg_rcmdr_biostat()

check_updates_rcmdr_biostat()

}}
}
\concept{packages}
