% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_update.R
\name{update_pkg_rcmdr_biostat}
\alias{update_pkg_rcmdr_biostat}
\title{Update package RcmdrPlugin.biostat and related packages}
\usage{
update_pkg_rcmdr_biostat()
}
\description{
Update package RcmdrPlugin.biostat and related packages
}
\examples{
\dontrun{\donttest{

update_pkg_rcmdr_biostat()

}}
}
