% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_update.R
\name{update_pkg_rcmdr_biostat}
\alias{update_pkg_rcmdr_biostat}
\title{Update package RcmdrPlugin.biostat and related packages}
\usage{
update_pkg_rcmdr_biostat(upgrade = TRUE, force = FALSE)
}
\arguments{
\item{upgrade}{Upgrade dependencies.
See \code{upgrade} in \code{\link[remotes:install_github]{remotes::install_github()}}.}

\item{force}{(logical) \code{TRUE} or \code{FALSE}.
Force to update.
See \code{force} in \code{\link[remotes:install_github]{remotes::install_github()}}.}
}
\description{
Update package RcmdrPlugin.biostat and related packages
}
\examples{
\dontrun{\donttest{

update_pkg_rcmdr_biostat()

}}
}
