% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_update.R
\name{update_pkg_snippets}
\alias{update_pkg_snippets}
\title{Update RcmdrPlugin.biostat and related packages}
\usage{
update_pkg_snippets()
}
\description{
Update RcmdrPlugin.biostat and related packages
}
\examples{
\dontrun{\donttest{

update_pkg_snippets()

}}
}
