% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_update.R
\name{update_pkg_snippets}
\alias{update_pkg_snippets}
\title{Update '\pkg{snippets}' and related packages}
\usage{
update_pkg_snippets(upgrade = FALSE, force = FALSE, quiet = TRUE)
}
\arguments{
\item{upgrade}{Upgrade dependencies.
See \code{upgrade} in \code{\link[remotes:install_cran]{remotes::install_cran()}}.}

\item{force}{(logical) \code{TRUE} or \code{FALSE}.
Force to update.
See \code{force} in \code{\link[remotes:install_cran]{remotes::install_cran()}}.}

\item{quiet}{(logical) \code{TRUE} or \code{FALSE}.
Should installation messages be printed?
See \code{quiet} in \code{\link[remotes:install_cran]{remotes::install_cran()}}.}
}
\description{
Update '\pkg{snippets}' and related packages
}
\examples{
\dontrun{\donttest{

update_pkg_snippets()

}}
}
\concept{packages}
