% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{version-control}
\alias{version-control}
\alias{open_rs_proj_vcs_opts}
\alias{show_rs_vcs_rsa_key}
\alias{browse_rs_vcs_help}
\title{Version control in RStudio.}
\usage{
open_rs_proj_vcs_opts()

show_rs_vcs_rsa_key()

browse_rs_vcs_help()
}
\description{
Helper functions for version control system (vcs) in RStudio.
\itemize{
\item \code{open_rs_proj_vcs_opts()} opens version control setup for RStudio project.
\item \code{show_rs_vcs_rsa_key()} shows version RSA key.
\item \code{browse_rs_vcs_help()} opens website with help of version control in RStudio .
}
}
